% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Archive0.R
\docType{class}
\name{Archive0}
\alias{Archive0}
\title{Archive0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Archive0
}
\description{
\code{Archive0} Singleton class for Archive subclass.
}
\details{
This class ...
}
\section{Methods}{

\describe{
 \item{\code{initialize(...)}}{Initializes a Singleton object}
 \item{\code{getInstance(...)}}{Gets instance of singleton object }
}
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
