% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Authorizer.R
\docType{class}
\name{Authorizer}
\alias{Authorizer}
\title{Authorizer}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Authorizer
}
\arguments{
\item{class}{Character string containing the name of valid class in the NLPStudio package.}

\item{attributes}{Vector containing the names of the attributes of the class.}

\item{subject}{Character string containing the name of the active class requesting access.}

\item{object}{Character string containing the name of the class to which access is being requested.}

\item{attribute}{Character string containing the name of the class's attribute to which access is being requested.}

\item{type}{Character representing the type of access being requested. Valid types are "r" for read and "w" for write access.}
}
\value{
A logical, TRUE if authorization is granted, FALSE otherwise.
}
\description{
\code{Authorizer} Class which maintains the authentication and access
policy and authorizes activities.
}
\details{
\strong{Authorizer Methods:}
\describe{
 \item{\code{new()}}{Instantiates the authorizer.}
 \item{\code{addClass(class, attributes)}}{Adds a class and its attributes to the list of valid classes and attributes}#'
 \item{\code{removeClass(class)}}{Removes a class from the Authorizer.}
 \item{\code{addAttribute(class, attribute)}}{Adds an attribute to a class.}
 \item{\code{removeAttribute(class)}}{Removes an attribute from a class .}
 \item{\code{addPolicy(subject, object, attribute, type)}}{Creates an authorization policy.}
 \item{\code{removePolicy(subject, object = null, attribute = null)}}{Removes a policy based upon the parameters presented. The parameters are validated from left to right. e.g. if attribute is sent, the object can not be null.}
 \item{\code{authorize(subject, object, attribute, type)}}{Returns true if the subject is authorized to perform the requested type of action on the attribute.}
}

\strong{Authorizer Parameters:}
}
\seealso{
Other Validation Classes: \code{\link{Validator0}},
  \code{\link{ValidatorClass}},
  \code{\link{ValidatorDate}},
  \code{\link{ValidatorExists}},
  \code{\link{ValidatorLogical}},
  \code{\link{ValidatorName}},
  \code{\link{ValidatorNoSpaces}},
  \code{\link{ValidatorNotEmpty}},
  \code{\link{ValidatorNumeric}},
  \code{\link{ValidatorPath}},
  \code{\link{ValidatorString}},
  \code{\link{ValidatorUrl}}, \code{\link{Validator}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
