% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildData.R
\docType{class}
\name{BuildData}
\alias{BuildData}
\title{BuildData}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BuildData
}
\arguments{
\item{name}{Character string including the name of the Corpus object}

\item{desc}{Character string including a description for the Corpus object}
}
\description{
\code{BuildData} Class responsible for building the processed data sets.
}
\details{
This class obtains the corpus from external sources, creates the Corpus and
Document class objects, performs text parsing, normalization, and filtering.

The Document family of classes is an implementation of the builder
pattern documented in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). This pattern allows the
construction process to be defined at runtime. The end product of
the Builder process is the Data Analytics Platform (DAP).
}
\section{Builder0 Participants}{

 \describe{
  \item{Builder0}{The abstract builder class, which defines a common interface and methods for the Builder family of classes.}
  \item{BuildData}{Class responsible for sourcing and initializing the Corpus and Document objects.}
  \item{BuildFeatures}{Class responsible for feature engineering, such as NGrams, POS tags, and TDM.}
  \item{BuildModelMKN}{Class responsible for building the MKN Language Model.}
  \item{BuildModelKatz}{Class responsible for building the Katz Language Model.}
  }
}

\section{Build Document Methods}{

 \describe{
  \item{\code{new()}}{Instantiates the builder object.}
  \item{\code{getData()}}{Obtains the corpus data from an external source.}
  \item{\code{BuildData()}}{Builds the document objects objects.}
  \item{\code{encodeDocuments()}}{Corrects file encoding.}
 }
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
