% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Command0.R
\docType{class}
\name{Command0}
\alias{Command0}
\title{Command0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Command0
}
\description{
\code{Command0} Abstract class for the command classes.
}
\details{
Abstract class for the command classes.
}
\section{Methods}{

\describe{
 \item{\code{execute()}}{Method that executes the command passed in from the Parser Class }
}
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
