% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\docType{class}
\name{Constants}
\alias{Constants}
\title{Constants}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Constants
}
\description{
\code{Constants} Class containing constants used throughout the package such
as directories structures.
}
\section{Constants Methods}{

\describe{
 \item{\code{getPaths()}}{Returns the NLPStudio directory and file paths as a list.}
}
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
