% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Corpus.R
\docType{class}
\name{Corpus}
\alias{Corpus}
\title{Corpus}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Corpus
}
\arguments{
\item{dataSource}{A DataSource or}

\item{field}{Character string name for a field to be added to the document or corpus object meta data.}

\item{name}{Character string containing the name of the corpus object. This variable is used in the instantiation and remove methods.}

\item{path}{Character string containing the path to the corpus object.}

\item{visitor}{An object of one of the visitor classes.}
}
\description{
\code{Corpus} Class of document collections or corpora.
}
\details{
Class that contains document collections, or corpora, as well as the methods
for reading, writing, and managing the documents therein.
}
\section{Methods}{

\strong{Core Methods}
 \describe{
  \item{\code{new(name, path, dataSource)}}{Instantiates an object of the Corpus class.}
 }

\strong{Document Methods}
 \describe{
  \item{\code{getDocuments()}}{Returns the Corpus Document objects.}
  \item{\code{addDocument(document)}}{Adds a Document object to the Corpus.}
  \item{\code{removeDocument(document)}}{Removes a Document object from the Corpus.}
 }

\strong{IO Methods}
 \describe{
  \item{\code{read(io = NULL)}}{Reads the corpus from file.}
  \item{\code{write(io = NULL)}}{Writes the corpus to file.}
 }

\strong{Metadata Methods}
 \describe{
  \item{\code{docMeta(field)}}{Creates a document meta data field.}
  \item{\code{corpusMeta(field)}}{Creates a corpus meta data field.}
  \item{\code{metaVarNames()}}{Prints Corpus and Document metadata variables.}
 }

\strong{Other Methods}
 \describe{
  \item{\code{logIt(level = 'Info')}}{Formats the log and calls the LogR class to log an event.}
  \item{\code{accept(visitor)}}{Accepts an object of the Visitor family of classes.}
 }
}

\section{Parameters}{
NA
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
