% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Corpus0.R
\docType{class}
\name{Corpus0}
\alias{Corpus0}
\title{Corpus0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Corpus0
}
\arguments{
\item{name}{A character string containing the name of the Corpus0 object. This variable is used in the instantiation and remove methods.}

\item{desc}{A chararacter string containing the description of the Corpus0.}

\item{document}{An object of one of the Document sub-classes.}

\item{visitor}{An object of one of the visitor classes.}
}
\description{
\code{Corpus0} Abstract class for the Corpus classes.
}
\details{
Class defines core functions common to the Corpus family of classes.
}
\section{Corpus0 Core Methods}{

 \describe{
  \item{\code{new(name, desc = NULL, studio = NULL)}}{Creates an object of Corpus0 Class}
  \item{\code{desc}}{A getter/corpuster method allowing clients to retrieve and corpus the Corpus0 description variable.}
  \item{\code{parent}}{A getter/corpuster method allowing clients to retrieve and corpus the Studio object to which the Corpus0 object belongs.}
  \item{\code{getName()}}{Returns the name of the Corpus0 object.}
  \item{\code{getPath()}}{Returns the path of the Corpus0 object.}
  \item{\code{getDocuments()}}{Returns the list of Document class objects.}
  \item{\code{addDocument(document)}}{Adds a Document object to the corpus.}
  \item{\code{removeDocument(document)}}{Removes a Document object from the corpus.}
  \item{\code{logIt(level = 'Info', fieldName = NA)}}{Formats the log and calls the LogR class to log an event.}
  \item{\code{accept(visitor)}}{Accepts an object of the Visitor family of classes.}
 }
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
