% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorpusBuilderCv.R
\docType{class}
\name{CorpusBuilderCV}
\alias{CorpusBuilderCV}
\title{CorpusBuilderCV}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CorpusBuilderCV
}
\arguments{
\item{name}{Character string indicating the name of the Corpus object}

\item{desc}{Character string containing the description of the Corpus}

\item{sourceCmd}{Command object containing the commands for sourcing the corpus.}

\item{parseCmd}{Command object containing the commands for parsing the corpus.}

\item{cvCmd}{Command object containing the commands for building the cross-validation sets.}
}
\description{
\code{CorpusBuilderCV} Concrete class for building the traditional training, validation and test cross-validation sets.
}
\details{
Class sources, repairs, and processes training, validation and test set corpora.

The CorpusBuilder family of classes is an implementation of the builder
pattern documented in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). This pattern allows the
construction process to be defined at runtime.
}
\section{Corpus Builder Participants}{

 \describe{
  \item{CorpusBuilder0}{This abstract builder interface. Defines the interface for concrete corpus builder sub-classes. }
  \item{CorpusBuilderCv}{Concrete builder sub-class that produces the a single training, validation, and test sets. }
  \item{CorpusBuilderKFoldCv}{Concrete builder sub-class that produces the K-Fold cross validation sets. }
  \item{CorpusDirector}{Class that builds the corpus through the concrete builder interfaces.}
  \item{Corpus}{The corpus product.}
  }
}

\section{Corpus CorpusBuilder0 Methods}{

 \describe{
  \item{\code{new(name, desc, sourceCmd, parseCmd, cvCmd)}}{Instantiates the builder object}
  \item{\code{buildRaw()}}{Obtains and stores the raw data.}
  \item{\code{buildRefined()}}{Builds the refined data sets with encoding errors corrected.}
  \item{\code{buildReshaped()}}{Builds a corpus reshaped into sentences.}
  \item{\code{buildCV()}}{Builds the cross validation sets.}
  \item{\code{buildParsed()}}{Builds the parsed data sets.}
  \item{\code{buildNormalized()}}{Builds the normalized data sets.}
 }
}

\seealso{
Other Corpus builder classes: \code{\link{CorpusBuilder0}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
