% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorpusBuilderRaw.R
\docType{class}
\name{CorpusBuilderRaw}
\alias{CorpusBuilderRaw}
\title{CorpusBuilderRaw}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CorpusBuilderRaw
}
\arguments{
\item{corpusName}{String containing the name of the corpus to be built.}

\item{corpusDesc}{String containing the dscription of the corpus to be built.}

\item{documentNames}{Character vector listing one or more names of the documents in the corpus.}

\item{documentFileNames}{Character vector listing one or more names for the files in the coprus source.}

\item{documentDescs}{Character vector listing one or more file descriptions for the documents in the corpus.}

\item{corpusSource}{Object of class CorpusSource, indicating the methods for sourcing the corpus.}

\item{analyzer}{Object of class Analyzer that conducts an analysis on the obtained corpus.}

\item{designer}{Object of class Designer that designates the features of the corpus.}

\item{processor}{Object of class Processor that renders the corpus.}

\item{url}{String containing the URL from which the corpus will be sourced}

\item{zipFile}{String containing the name of the zipfile to be downloaded}
}
\value{
An object of the DocumentCollection class and one instance of the Document class
for each file or register within the corpus.  The objects will contain the corpus
content and  meta data. The content will also be stored on disk.
}
\description{
\code{CorpusBuilderRaw} Concrete builder class for obtaining and preparing raw corpora
}
\details{
\strong{CorpusBuilder Class Overview:}

The CorpusBuilder family of classes is an implementation of the builder
pattern documented in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). This pattern allows for different
representations of a corpus to be constructed, independent of the
components that comprise the corpus. Corpora are constructed to  support
various stages of analysis from raw, preprocessed, clean, cross-validation
and language modeling.
}
\section{CorpusBuilder Class Participants}{

The participants of the CorpusBuilder
class are:
\enumerate{
 \item CorpusBuilder0: Specifies an abstract interface for creating parts
 of the composite Document0 Class object.
 \item CorpusDirector: Constructs the objects using the CorpusBuilder0
 interface.
 \item CorpusBuilderRaw: Constructs a raw corpus originating from a range
 of sources including web and directory sources.
 \item CorpusBuilderPreprocessor: Constructs a preprocessed corpus
 from the raw corpus. This will include basic encoding fixes and tokenization.
 \item CorpusBuilderClean: Constructs a clean corpus. Interacts with the
 CorpusDesigner0 and CorpusProcessor0 classes to design and render the
 clean corpus.
 \item CorpusBuilderPilot. Constructs a representative sample of the
 clean corpus for exploratory data analysis, modeling and evaluation. This
 method interfaces with the CorpusAnalyzer0, CorpusDesigner0, and
 CorpusProcessor0 classes to analyze the originating corpus, design
 the pilot corpus and process.
 \item CorpusBuilderValidationSets. Constructs training, validation
 and test sets.  Interacts with the CorpusAnalyzer0, CorpusDesigner0,
 and CorpusProcessor0 classes to analyze, design, and render the
 training, validation, and test set objects.
 }
}

\section{CorpusBuilder Generic Interface}{

The abstract interface in the CorpusBuilder0 class outlines "generic"
algorithm common to all concrete builder classes. The core methods are
defined as follows:
\enumerate{
 \item initialize: Instantiates an object of a concrete CorpusBuilder
  classes.
 \item buildCollection: Constructs an object of the DocumentCollection
 class. This is a composite class containing other DocumentCollection
 objects or individual objects of the Document class.
 \item buildDocument: Constructs an object of the Document class. This
 method requires the designation of file names as a required parameter.
 \item sourceCorpus: Obtains the originator corpus from sources such
 as web sites or directories, and stores it to disk.
 \item analyzeCorpus: Analyzes the corpus via the CorpusAnalyzer0
 interface, an implementation of the strategy pattern.  Data quality,
 word frequency, and semantic analyses are a few of the types of analyses
 supported.
 \item designCorpus: Interacts with the CorpusDesigner0 interface,
 allowing the client to design a corpus based upon features to be
 included/excluded, sample sizes, etc...
 \item processCorpus: Constructs the corpus using the CorpusProcessor0
 interface.
 \item evaluateCorpus: Interfaces with the CorpusAnalyzer0 class to
 confirm that the corpus conforms to designated design features.
}
}

\section{CorpusBuilder Collaborations}{

The sourcing, analysis, design, and processing steps are implemented using
various strategy patterns. The following behavioral methods have been
devised to select concrete implementations of the strategy classes and are
summarized as follows:
\itemize{
 \item getCorpusSource: Method for selecting the corpus source method.
 \item setCorpusSource:  Method for setting the corpus source methods.
 \item getAnalyzer: Method for selecting the analysis methods
 \item setAnalyzer: Method for setting the analysis methods.
 \item getDesigner: Method for selecting the corpus designer method.
 \item setDesigner: Method for setting the corpus designer methods.
 \item getProcessor: Method for selecting the corpus processor method.
 \item setProcessor: Method for setting the corpus processor methods.
}
}

\section{CorpusBuilderRaw Interface}{

The concrete implementation of the abstract interface includes both core and behavioral interfaces.

\strong{Core interfaces:}
\describe{
 \item{\code{new(name)}}{Instantiates object of CorpusBuilderRaw class.}
 \item{\code{getBuilder()}}{Returns values of CorpusBuilderRaw object}
 \item{\code{buildCollection(corpusName, corpusDesc = NULL)}}{Instantiates object of DocumentCollection class.}
 \item{\code{buildCDocument(documentNames, fileNames)}}{Instantiates object of Document class. Both parameters are required.}
 \item{\code{sourceCorpus(url, downloadFile)}}{Downloads the corpus from the designated URL and unzips the content into the downloadFile.  The files are stored in the directory designated for this object and data are loaded into the Document objects in memory.}
 \item{\code{analyzeCorpus()}}{Abstract method not implemented in this class.}
 \item{\code{designCorpus()}}{Abstract method not implemented in this class.}
 \item{\code{processCorpus()}}{Abstract method not implemented in this class.}
 \item{\code{evaluateCorpus()}}{Abstract method not implemented in this class.}
 \item{\code{getCorpus()}}{Returns the DocumentCollection object and a Document object for each file in the DocumentCollection.}
}

\strong{Behavior Interfaces:}
\describe{
 \item{\code{corpusSource(value)}}{Method for getting and setting the corpus sources.}
 \item{\code{analyzer(value)}}{Not implemented in this class.}
 \item{\code{designer(value)}}{Not implemented in this class.}
 \item{\code{processor(value)}}{Not implemented in this class.}
}
}

\seealso{
Other CorpusBuilder Classes: \code{\link{CorpusBuilder0}},
  \code{\link{CorpusDirector}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
