% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorpusDirector.R
\docType{class}
\name{CorpusDirector}
\alias{CorpusDirector}
\title{CorpusDirector}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CorpusDirector
}
\description{
\code{CorpusDirector} Director class of the CorpusBuilder set of classes
}
\details{
This director class is part of the CorpusBuilder set of classes, an
implementation of the builder design pattern. The class interacts with
the builder interfaces to build a document corpus for use throughout
the data acquisition, cleaning, analysis, and modeling phases. The
standard interface methods obtain and process the corpus,
create the composite document and return the corpus meta data and
content to the calling environment.
}
\seealso{
Other CorpusBuilder Classes: \code{\link{CorpusBuilder0}},
  \code{\link{CorpusBuilderRaw}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
