% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorpusSourceWeb.R
\docType{class}
\name{CorpusSourceWeb}
\alias{CorpusSourceWeb}
\title{CorpusSourceWeb}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CorpusSourceWeb
}
\description{
\code{CorpusSourceWeb} Class responsible for obtaining corpus data from web sources.
}
\details{
Provides the basic interface and core methods for the CorpusSource
concrete classes.

The CorpusSourceWeb family of classes is an implementation of the strategy
pattern documented in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). This pattern allows corpus
sourcing behavior to be defined at run time.
}
\section{CorpusSourceWeb Family Participants}{

 \itemize{
  \item{CorpusSource0}{Abstract class that defines core methods.}
  \item{CorpusSourceWeb}{Class responsible for obtaining corpus data from web sources.}
 }
}

\section{CorpusSourceWeb method}{

 \itemize{
  \item{\code{new(...)}}{Method not implemented for this abstract class.}
  \item{\code{sourceData()}}{Method not implemented for this abstract class.}
 }
}

\seealso{
Other CorpusSource classes: \code{\link{CorpusSource0}},
  \code{\link{CorpusSource}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
