% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Curator.R
\docType{class}
\name{Curator}
\alias{Curator}
\title{Curator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Curator
}
\arguments{
\item{className}{Character string containing the class to which the query objects belong.}

\item{objectName}{Character string containing the name of the object being queried.}

\item{dateFrom}{ISO or 8601 formatted date indicating the date from which data should be returned.}

\item{dateTo}{ISO or 8601 formatted date indicating the date to which data should be returned.}

\item{lastN}{Numeric indicating the last N minutes, hours, days weeks or months for which data should be returned. Values will be rounded to nearest integer.}

\item{unit}{Character string indicating the unit of time associated with the lastN variable. Valid values include c("min", "hour", "day", "week", "month")}
}
\description{
\code{Curator} Class responsible for building, validating and executing
queries into objects that manage history or states of other objects.
Concretely, this class manages queries into Historian and StateManager
objects.
}
\details{
\strong{Curator Methods:}
The Curator methods are as follows:
 \itemize{
  \item{\code{initialize(target)}}{Initializes a Curator object for the target object.}
  \item{\code{buildQuery()}}{Builds the query for the target object.}
  \item{\code{submitQuery()}}{Submits the query to the target object.}
  \item{\code{getQuery()}}{Returns the query to the calling method.}
}

\strong{Getter/Setter Methods:}
The query can be built, parameter at a time with the following getter/setter methods
 \itemize{
  \item{\code{className}}{Getter/setter method for the classNsme query parameter.}
  \item{\code{objectName}}{Getter/setter method for the objectName query parameter.}
  \item{\code{dateFrom}}{Getter/setter method for the dateFrom query parameter.}
  \item{\code{dateTo}}{Getter/setter method for the dateTo query parameter.}
  \item{\code{lastN}}{Getter/setter method for the lastN query parameter.}
  \item{\code{unit}}{Getter/setter method for the unit query parameter.}
}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
