% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{class}
\name{Data}
\alias{Data}
\title{Data}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Data
}
\arguments{
\item{corpus}{Corpus object.}
}
\description{
\code{Data} Class which contains all corpora and cross validation sets.
}
\details{
Class containing the raw, refined, cross-validation and preprocessed corpora.
The cross-validation sets contain a single training, validation, and
test set.
}
\section{Data Methods}{

 \describe{
  \item{\code{new(train, val, test)}}{Instantiates the Data object}
  \item{\code{getCorpora()}}{Retrieves the list of corpora from the data set.}
  \item{\code{addCorpus(corpus)}}{Adds a corpus to the data set.}
  \item{\code{removeCorpus(corpus))}}{Removes a corpus from the data set.}
  \item{\code{move(parent))}}{Moves the data set to the new parent object.}
 }
}

\seealso{
Other Data builder classes: \code{\link{CorpusBuilder0}},
  \code{\link{DataBuilder}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
