% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document.R
\docType{class}
\name{Document}
\alias{Document}
\title{Document}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Document
}
\arguments{
\item{filePath}{Character string indicating the file path for a document}

\item{parent}{Object of the Corpus or Set classes to which this document belongs}

\item{io}{An object of one of the IO classes used for reading and writing.}
}
\description{
\code{Document} Class that contains document meta data and text content.
}
\details{
This class contains the data and methods for creating, reading, manipulating,
and processing text documents.
}
\section{Document core methods}{

 \itemize{
  \item{\code{new(filePath, desc = NULL)}}{Method for instantiating a document.}
  \item{\code{getName()}}{Method that returns the name of the Document object.}
  \item{\code{getFileName()}}{Method for obtaining the document file name.}
  \item{\code{getPath()}}{Method for obtaining the document path.}
  \item{\code{getContent()}}{Method for obtaining the document content.}
 }
}

\section{Document getter/setter methods}{

 \itemize{
  \item{\code{desc()}}{Method for setting or retrieving the Document object description.}
 }

 @section Document IO methods:
 \itemize{
  \item{\code{addContent(content)}}{Method for adding content to a document.}
  \item{\code{read(io)}}{Method for reading a document.}
  \item{\code{write(io)}}{Method for writing a document.}
 }
}

\section{Document aggregation method}{

 \itemize{
  \item{\code{move(parent)}}{Moves a document to the designated parent corpus.}
 }

 @section Document Processing Methods:
 \itemize{
  \item{\code{refine()}}{Repairs encoding and format errors in text document.}
  \item{\code{reshapeSent()}}{Reshapes the document into sentences.}
  \item{\code{parseFast()}}{Relatively fast text parser.}
  \item{\code{parseEmail()}}{Parses email addresses from the document.}
  \item{\code{parseControl()}}{Parses control characters from the document.}
  \item{\code{parseRepeatChars()}}{Parses repeated characters from the document.}
  \item{\code{parseLongWords()}}{Parses words > 40 chars, from the document.}
  \item{\code{normalize()}}{Normalizes abbreviations and contractions in a document.}
  \item{\code{sanitizeWord()}}{Removes profane words from a document.}
  \item{\code{sanitizeSent()}}{Removes sentences containing profane words from a document.}
  \item{\code{sanitizeTag()}}{Replaces profane words with the <EXPLETIVE> tag.}
  \item{\code{cleanUp()}}{Removes extra white-space, stray hyphens an empty sentences.}
 }
}

\section{Other methods}{

 \itemize{
  \item{\code{accept(visitor)}}{Accepts a visitor object.}
  \item{\code{logIt(level = 'Info')}}{Logs events relating to the document.}
 }
}

\seealso{
Other Document classes: \code{\link{Preprocess}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
