% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentSource.R
\docType{class}
\name{DocumentSource}
\alias{DocumentSource}
\title{DocumentSource}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DocumentSource
}
\arguments{
\item{downloadParams}{A list containing the following fields:
\itemize{
 \item{url}{The URL from which the content will be downloaded.}
 \item{destFile}{The file path into which the data will be downloaded.}
}}

\item{zipParams}{A list containing the following fields:
\itemize{
 \item{zipFiles}{A character vector including the files to be extracted from the zip file.}
 \item{unZipPath}{A character string containing the directory into which the zip files are to be extracted.}
}}

\item{source}{A character string containing the instantiating method for the sourcing strategy.  At present, there are two: WebSource and CollectionSource.}
}
\description{
\code{DocumentSource} Class for obtaining documents sourced from the internet, other document collections,
}
\details{
This leaf class is part of the Document0 composition class. This class
includes the data and methods required to source a document collection from
web source or other collections. The client instantiates an object of the
DocumentSource class, sets the sourcing strategy, then gets the document by
executing the sourcing strategy.  The current strategies are web and collection.
}
\section{Methods}{

\describe{
 \item{\code{new()}}{Creates an object of the DocumentSource class}
 \item{\code{source}}{Returns the current document source strategy}
 \item{\code{source <- value}}{Sets the current document source strategy. Current options are WebSource$new() and CollectionSource$new()}
 \item{\code{getDocument}}{Implements the document source strategy}
}
}

\section{Parameters}{

}

\examples{
\dontrun{
  webCorpus <- DocumentSource$new()
  webCorpus$source <- WebSource$new()
  webCorpus$getDocument(downloadParams, zipParams)
}

}
\seealso{
Other Document classes: \code{\link{Document0}},
  \code{\link{DocumentCollection}}, \code{\link{Document}}
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
\keyword{datasets}
