% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentText.R
\docType{class}
\name{DocumentText}
\alias{DocumentText}
\title{DocumentText}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DocumentText
}
\arguments{
\item{filePath}{Character string indicating the file path for a document}

\item{parent}{Object of the Corpus or Set classes to which this document belongs}

\item{io}{An object of one of the IO classes used for reading and writing.}

\item{repairs}{A list of key value pairs of strings to be replaced in the repair step}

\item{norms}{A list of key value pairs of strings to be replaced in the normalization step}

\item{corrections}{A list of key value pairs of strings to be replaced in the corrections step}

\item{profanity}{Character vector of profane words to be extracted from the text}
}
\description{
\code{DocumentText} Class that contains document meta data and text content.
}
\details{
This class contains the data and methods for creating, reading, manipulating,
processing, and transforming text documents.
}
\section{Document Family Participants}{

 \itemize{
  \item{Document0}{This abstract class defines the interface for the other classes.}
  \item{DocumentText}{Class for text documents.}
  \item{DocumentNGrams}{Class for NGrams.}
  \item{DocumentPOSTags}{Class for POS tags.}
 }
}

\section{Document core methods}{

 \itemize{
  \item{\code{new(filePath, desc = NULL)}}{Method for instantiating a document.}
  \item{\code{getName()}}{Method that returns the name of the Document object.}
  \item{\code{getFileName()}}{Method for obtaining the document file name.}
  \item{\code{getPath()}}{Method for obtaining the document path.}
  \item{\code{readDocument(io)}}{Method for initiating the read operation for a document.}
  \item{\code{writeDocument(io)}}{Method for initiating the write operation for a document.}#'
  \item{\code{repairDocument(repairs)}}{Method for repairing a document.}
  \item{\code{normalizeDocument(norms)}}{Method for normalizing text in a document.}
  \item{\code{sanitizeDocument(profanity)}}{Method for removing profanity from a text document.}
 }
}

\seealso{
Other Document classes: \code{\link{Document0}},
  \code{\link{Preprocess}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
