% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentTxt.R
\docType{class}
\name{DocumentTXT}
\alias{DocumentTXT}
\title{DocumentTXT}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DocumentTXT
}
\arguments{
\item{parent}{Parent Corpus object}

\item{name}{Character string containing the name of the document}

\item{desc}{Character string containing the description of the document}

\item{path}{Character string containing the path}
}
\description{
\code{DocumentTXT} Class that contains document meta data and text content for a txt file
}
\details{
This class contains the data and methods for creating, reading, manipulating,
and processing text documents.
}
\section{DocumentTXT core methods}{

 \itemize{
  \item{\code{new(filePath, desc = NULL)}}{Method for instantiating a document.}
  \item{\code{getName()}}{Method that returns the name of the DocumentTXT object.}
  \item{\code{getFileName()}}{Method for obtaining the document file name.}
  \item{\code{getPath()}}{Method for obtaining the document path.}
 }
}

\section{DocumentTXT getter/setter methods}{

 \itemize{
  \item{\code{desc()}}{Method for setting or retrieving the DocumentTXT object description.}
 }

 @section DocumentTXT IO methods:
 \itemize{
  \item{\code{getContent()}}{Method for obtaining content from a document.}
  \item{\code{addContent(content)}}{Method for adding content to a document.}
  \item{\code{read(io)}}{Method for reading a document.}
  \item{\code{write(io)}}{Method for writing a document.}
 }
}

\section{DocumentTXT aggregation method}{

 \itemize{
  \item{\code{setParent(parent)}}{Creates a reference to the parent object. }
 }
}

\section{Other methods}{

 \itemize{
  \item{\code{accept(visitor)}}{Accepts a visitor object.}
  \item{\code{logIt(level = 'Info')}}{Logs events relating to the document.}
 }
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
