% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download0.R
\docType{class}
\name{Download0}
\alias{Download0}
\title{Download0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Download0
}
\description{
\code{Download0} Abstract class for downloading corpora from the internet
}
\details{
This abstract class encapsulates the behaviors used to download corpora
from internet sources.
}
\section{Methods}{

\describe{
 \item{\code{download()}}{Is overwritten by concrete subclasses}
}
}

\author{
John James, \email{j2sdatalab@gmail.com}
}
\keyword{datasets}
