% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadBinary.R
\docType{class}
\name{DownloadBinary}
\alias{DownloadBinary}
\title{DownloadBinary}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DownloadBinary
}
\description{
\code{DownloadBinary} Class for downloading a corpus from the internet using binary mode.
}
\details{
This class includes the data and methods required to download a corpus from
an internet source using binary mode.
}
\section{Methods}{

\describe{
 \item{\code{download()}}{Downloads the corpus from the internet source to the designated destfile}
}
}

\author{
John James, \email{j2sdatalab@gmail.com}
}
\keyword{datasets}
