% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entity.R
\docType{class}
\name{Entity}
\alias{Entity}
\title{Entity}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Entity
}
\description{
\code{Entity} Base class for all entity related classes
}
\details{
This base class defines members and methods common across all entity related
classes, including Studio, Corpus, Set, and Document classes.
}
\section{Document methods}{

 \itemize{
  \item{\code{desc()}}{Active binding getter/setter for object description.}
  \item{\code{getName()}}{Method for retrieving an object's name.}
  \item{\code{getClassName()}}{Method for retrieving an object's class name}
  \item{\code{getPath()}}{Method for retrieving an object's path.}
 }
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
