% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Environment.R
\docType{class}
\name{Environment}
\alias{Environment}
\title{Environment}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Environment
}
\arguments{
\item{parent}{A character string containing the name of the parent environment.  The topmost environment is the nlpStudio object.}

\item{name}{A character string containing the name of the environment}

\item{desc}{A chararacter string containing the description of the environment}
}
\description{
\code{Environment} Leaf class for creating and printing environments
}
\details{
The Environment class, inherits the Environment0 abstract class and provides
methods for instantiating and printing objects of the Environment class.
}
\section{Methods}{

\describe{
 \item{\code{new()}}{Creates an object of environment "leaf" subclass}
 \item{\code{getEnv(verbose = FALSE)}}{Retrieves the meta data for the environment, as well as its corpora and states, returns the data in a list and, if verbose == TRUE, prints the data to the console}
}
}

\section{Fields}{

}

\author{
John James, \email{j2sdatalab@gmail.com}
}
\keyword{datasets}
