% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Environment0.R
\docType{class}
\name{Environment0}
\alias{Environment0}
\title{Environment0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Environment0
}
\arguments{
\item{parent}{A character string containing the name of parent the environment. The topmost environment is nlpStudio object.}

\item{name}{A character string containing the name of the environment}

\item{desc}{A chararacter string containing the description of the environment}
}
\description{
\code{Environment0} Abstract class for environment composite classes
}
\details{
The Environment0 class is an abstract class that defines the common and
composite tree methods for managing environments. The NLPStudio makes up
the single composite environment containing the child environments.
}
\section{Methods}{

\describe{
 \item{\code{new()}}{Creates an object of environment "leaf" subclass}
 \item{\code{getEnv(verbose = FALSE)}}{Retrieves the meta data for the environment, as well as its corpora and states, returns the data in a list and, if verbose == TRUE, prints the data to the console}
  \item{\code{addEnv()}}{Adds the environment to the NLPStudio list of environments}
  \item{\code{archiveEnv()}}{Archives the files contained in an environment}
  \item{\code{removeEnv()}}{Archives and removes the files contained in an environment}
  \item{\code{listEnvs()}}{Lists the environments within the NLPStudio}
  \item{\code{searchEnvs(envName)}}{Search the environments by the environment name variable envName}
}
}

\section{Fields}{

}

\author{
John James, \email{j2sdatalab@gmail.com}
}
\keyword{datasets}
