% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileManager.R
\docType{class}
\name{FileManager}
\alias{FileManager}
\title{FileManager}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileManager
}
\arguments{
\item{name}{A character string containing the name of the FileManager object. This variable is used in the instantiation and remove methods.}

\item{desc}{A chararacter string containing the description of the FileManager}

\item{document}{An object of the DocumentCollection class to be added to the FileManager object's list of document collections.}

\item{visitor}{An object of one of the visitor classes.}
}
\description{
\code{FileManager} Class containing file management functions
}
\details{
This class contains functions for maintaining and manipulating files
}
\section{FileManager Methods}{

 \describe{
  \item{\code{new()}}{Creates an object of FileManager Class}
  \item{\code{move(from, to)}}{Move a file or directory from directory a to directory b.}
  \item{\code{copy(from, to)}}{Copy a file or directory from directory a to directory b.}
  \item{\code{zip(zipFile, files)}}{Compresses files or files and stores them in the filepath indicated by zipFile.}
  \item{\code{unzip(zipFile, exDir)}}{Extract compressed files and places them in the directory indicated by exDir.}
  \item{\code{download(url, destFile)}}{Downloads a file and stores it in .}
  \item{\code{checkDir(directory)}}{Creates a directory, if it already exists, it returns the size of the directory.}
 }
}

\section{Parameters}{

}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
