% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\docType{class}
\name{IO}
\alias{IO}
\title{IO}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
IO
}
\arguments{
\item{document}{Object of the Document family of classes}
}
\description{
\code{IO} Class responsible for reading and writing xlsx, csv, text, and rdata files through the read and write visitor methods.
}
\details{
\strong{IO Class Overview:}
The IO class is an implementation of the visitor design pattern,
as described in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). This visitor pattern allows
new operations to be defined without changing the classes upon which
the visitor method operates.

\strong{IO Methods:}
The IO class supports csv, Rdata, and text files through the following methods:
 \itemize{
  \item{\code{read(document)}}{Read method.}
  \item{\code{write(document)}}{Write method.}
}
}
\seealso{
Other Read/Write Classes: \code{\link{VReader}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
