% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO0.R
\docType{class}
\name{IO0}
\alias{IO0}
\title{IO0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
IO0
}
\arguments{
\item{document}{Object of the Document family of classes}
}
\description{
\code{IO0} Abstract class that defines the interface for reading and writing documents.
}
\details{
\strong{IO Class Overview:}
The IO0 class is an implementation of the strategy design pattern,
as described in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). This strategy pattern accommodates
various formats and allows the behavior to be defined / selected at run time.

\strong{IO0 Methods:}
The IO0 class supports csv, Rdata, and text files through the following methods:
 \itemize{
  \item{\code{read(document)}}{Read method.}
  \item{\code{write(document)}}{Write method.}
}
}
\seealso{
Other Input / Output Classes: \code{\link{IOBin}},
  \code{\link{IOCSV}}, \code{\link{IORDS}},
  \code{\link{IORdata}}, \code{\link{IOText}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
