% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOArchive.R
\docType{class}
\name{IOArchive}
\alias{IOArchive}
\title{IOArchive}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
IOArchive
}
\arguments{
\item{document}{Object of the Document family of classes}
}
\description{
\code{IOArchive} Class responsible for reading and writing Archive documents.
}
\details{
\strong{IO Class Overview:}
The IOArchive class is an implementation of the strategy design pattern,
as described in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). This strategy pattern accommodates
various formats and allows the behavior to be defined / selected at run time.

\strong{IOArchive Methods:}
The IOArchive class supports csv, Archive, and text files through the following methods:
 \itemize{
  \item{\code{read(document)}}{Read method.}
  \item{\code{write(document)}}{Write method.}
}
}
\seealso{
Other Input / Output Classes: \code{\link{IO0}},
  \code{\link{IOBin}}, \code{\link{IOCSV}},
  \code{\link{IORdata}}, \code{\link{IOText}},
  \code{\link{IOXlsx}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
