% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOCSV.R
\docType{class}
\name{IOCSV}
\alias{IOCSV}
\title{IOCSV}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
IOCSV
}
\arguments{
\item{file}{Object of the File family of classes}
}
\description{
\code{IOCSV} Class responsible for reading and writing csv files.
}
\details{
\strong{IO Class Overview:}
The IOCSV class is an implementation of the strategy design pattern,
as described in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). This strategy pattern accommodates
various formats and allows the behavior to be defined / selected at run time.

\strong{IOCSV Methods:}
The IOCSV class supports csv, Rdata, and text files through the following methods:
 \itemize{
  \item{\code{read(file)}}{Read method.}
  \item{\code{write(file)}}{Write method.}
}
}
\seealso{
Other Input / Output Classes: \code{\link{IO0}},
  \code{\link{IOBin}}, \code{\link{IORdata}},
  \code{\link{IOText}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
