% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOFactory.R
\docType{class}
\name{IOFactory}
\alias{IOFactory}
\title{IOFactory}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
IOFactory
}
\arguments{
\item{fileName}{Character string containing the name of the file}
}
\value{
File class object
}
\description{
\code{IOFactory} Factory class that creates the appropriate IO class.
}
\details{
Class creates IO objects based upon the file type
}
\section{IOFactory methods}{

\describe{
 \item{\code{new()}}{Instantiates the factory.}
 \item{\code{getFileStrategy(fileName)}}{Instantiates and returns the appropriate File Strategy object.}
}
}

\section{Validation Class Parameters}{
NA
}

\seealso{
Other Validation Classes: \code{\link{Validator0}},
  \code{\link{ValidatorClass}},
  \code{\link{ValidatorDate}},
  \code{\link{ValidatorExists}},
  \code{\link{ValidatorLogical}},
  \code{\link{ValidatorNoSpaces}},
  \code{\link{ValidatorNotEmpty}},
  \code{\link{ValidatorNumeric}},
  \code{\link{ValidatorPath}},
  \code{\link{ValidatorString}},
  \code{\link{ValidatorUrl}}, \code{\link{Validator}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
