% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IORds.R
\docType{class}
\name{IORDS}
\alias{IORDS}
\title{IORDS}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
IORDS
}
\arguments{
\item{file}{Object of the File family of classes}
}
\description{
\code{IORDS} Class responsible for reading and writing rdata files.
}
\details{
\strong{IO Class Overview:}
The IORDS class is an implementation of the strategy design pattern,
as described in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). This strategy pattern accommodates
various formats and allows the behavior to be defined / selected at run time.

\strong{IORDS Methods:}
The IORDS class supports csv, RDS, and text files through the following methods:
 \itemize{
  \item{\code{read(file)}}{Read method.}
  \item{\code{write(file)}}{Write method.}
}
}
\seealso{
Other Input / Output Classes: \code{\link{IO0}},
  \code{\link{IOBin}}, \code{\link{IOCSV}},
  \code{\link{IORdata}}, \code{\link{IOText}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
