% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Korpus.R
\docType{class}
\name{Korpus}
\alias{Korpus}
\title{Korpus}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Korpus
}
\arguments{
\item{lab}{Object of class Lab to which the Corpus is a member.}

\item{name}{A character string containing the name of the Korpus object. This variable is used in the instantiation and remove methods.}

\item{desc}{A chararacter string containing the description of the Korpus.}

\item{document}{An object of one of the Document sub-classes.}

\item{visitor}{An object of one of the visitor classes.}
}
\description{
\code{Korpus} Class that defines a corpus or collection of documents
}
\details{
The class one or several documents in text form, as well as various
transformations such as nGrams and POS tags
}
\section{Korpus Core Methods}{

 \describe{
  \item{\code{new(name, desc = NULL, lab = NULL)}}{Creates an object of Korpus Class}
  \item{\code{desc}}{A getter/setter method allowing clients to retrieve and set the Korpus description variable.}
  \item{\code{getName()}}{Returns the name of the Korpus object.}
  \item{\code{getPath()}}{Returns the path of the Korpus object.}
  \item{\code{getLogs()}}{Returns the LogR object for the Korpus object.}
  \item{\code{logIt(level = 'Info', fieldName = NA)}}{Formats the log and calls the LogR class to log an event.}
  \item{\code{accept(visitor)}}{Accepts an object of the Visitor family of classes.}
 }
}

\section{Korpus Aggregate Methods}{

 \describe{
  \item{\code{getDocuments()}}{Retrieves the list of documents for the Korpus.}
  \item{\code{addChild(document)}}{Adds a document to the Korpus object.}
  \item{\code{removeChild(document)}}{Removes a document from the Korpus object.}
}
}

\section{Korpus Read/Write Methods}{

 \describe{
  \item{\code{readKorpus()}}{Reads a corpus into memory .}
  \item{\code{writeKorpus()}}{Writes the contents of each document in the corpus to file.}
}
}

\section{Korpus Processing Methods}{

 \describe{
  \item{\code{repairKorpus()}}{Method for initiating the repair operation on a document.}
  \item{\code{cleanKorpus()}}{Method for cleaning the document.}
  \item{\code{nGramKorpus(nGrams)}}{Method for creating an nGram representation of the document.}
  \item{\code{posTagKorpus(nGrams)}}{Method for creating an POS tagged representation of the document.}
}
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
