% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KorpusBuilder.R
\docType{class}
\name{KorpusBuilder}
\alias{KorpusBuilder}
\title{KorpusBuilder0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
KorpusBuilder
}
\arguments{
\item{getData}{Boolean, indicator to obtain the data from an external source. Default is TRUE}

\item{documents}{Boolean, indicates whether document objects are to be created. Default is TRUE}

\item{repair}{Boolean, indicates whether the documents should undergo repair. Default is TRUE}

\item{split}{Boolean, indicates whether documents splits should be created. Default is TRUE}

\item{sample}{Boolean, indicates whether document samples should be taken. Default is TRUE}

\item{normalize}{Boolean, indicates whether text normalization should take place.Default is TRUE}

\item{correct}{Boolean, to perform corrections. Default is TRUE}

\item{profanity}{Boolean, indicator to remove profanity. Default is TRUE}

\item{nGrams}{Numeric indicator of the number of nGrams to build}

\item{posTags}{Boolean, indicates whether to create POS tags.  Default is TRUE}
}
\description{
\code{KorpusBuilder0} Abstract class for the Korpus builder classes
}
\details{
The Document family of classes is an implementation of the builder
pattern documented in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). This pattern allows the
construction process to be defined at runtime.
}
\section{Korpus Builder Participants}{

 \describe{
  \item{KorpusBuilder0}{This abstract builder interface. Defines the interface for concrete corpus builder sub-classes. }
  \item{KorpusDirector}{Class that builds the corpus through the concrete builder interface.}
  \item{Korpus}{The corpus or corpus set product.}
  }
}

\section{Korpus Builder0 Methods}{

 \describe{
  \item{\code{new()}}{Instantiates the builder object}
  \item{\code{getData()}}{Obtains the corpus data from an external source.}
  \item{\code{buildDocuments()}}{Builds document objects.}
  \item{\code{repairDocuments()}}{Rapairs corpus document objects}
  \item{\code{splitDocuments()}}{Creates random sample of 10 indices, each representing 10 pct of the lines of the document..}
  \item{\code{sampleDocuments()}}{Creates sammples of the document according to size increments of 10 pct.}
  \item{\code{normalizeDocuments()}}{Normalizes text in the document objects.}
  \item{\code{correctDocuments()}}{Corrects common misspellings, contractions, and abbreviations.}
  \item{\code{cleanLanguage()}}{Removes sentences containing profane language.}
  \item{\code{nGramDocuments()}}{Creates nGram representations of document objects.}
  \item{\code{posTagDocuments()}}{Creates POS tag representations of document objects.}
  \item{\code{deliverKorpus()}}{Returns Korpus object.}
 }
}

\seealso{
Other Corpus build family: \code{\link{KorpusBuilder0}},
  \code{\link{KorpusDirector}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
