% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KorpusBuilder0.R
\docType{class}
\name{KorpusBuilder0}
\alias{KorpusBuilder0}
\title{KorpusBuilder0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
KorpusBuilder0
}
\arguments{
\item{name}{Character string indicating the name of the Korpus object}

\item{desc}{Character string containing the description for the corpus object}

\item{parent}{Object of the Korpus or Lab classes}

\item{path}{Character string indicating the path to the Korpus object}

\item{dirs}{Directory structure for a Korpus object}

\item{state}{Character string indicating last state of object}

\item{url}{Character string indicating URL for downloading data}

\item{files}{Character vector containing names of files to be extracted from zipped archive.}

\item{repairs}{list of gsub key and value pairs}

\item{sets}{List of Document class objects, including text, nGrams and pos tags.}
}
\description{
\code{KorpusBuilder0} Abstract class for the Korpus builder classes
}
\details{
The Document family of classes is an implementation of the builder
pattern documented in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). This pattern allows the
construction process to be defined at runtime.
}
\section{Korpus Builder Participants}{

 \describe{
  \item{KorpusBuilder0}{This abstract builder interface. Defines the interface for concrete corpus builder sub-classes. }
  \item{KorpusBuilder}{Concrete builder sub-class that produces the Korpus object. }
  \item{KorpusSetBuilder}{Concrete builder sub-class that produces the Korpus training, validation and test set objects. }
  \item{KorpusDirector}{Class that builds the corpus through the concrete builder interfaces.}
  \item{Korpus}{The corpus or corpus set product.}
  }
}

\section{Korpus Builder0 Methods}{

 \describe{
  \item{\code{new()}}{Instantiates the builder object}
  \item{\code{getData()}}{Obtains the corpus data from an external source.}
  \item{\code{buildDocuments()}}{Builds document objects.}
  \item{\code{repairDocuments()}}{Rapairs corpus document objects}
  \item{\code{splitDocuments()}}{Creates random sample of 10 indices, each representing 10 pct of the lines of the document.}
 }
}

\seealso{
Other Corpus build family: \code{\link{KorpusBuilder}},
  \code{\link{KorpusDirector}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
