% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KorpusDirector.R
\docType{class}
\name{KorpusDirector}
\alias{KorpusDirector}
\title{KorpusDirector}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
KorpusDirector
}
\arguments{
\item{builder}{Corpus builder object}
}
\description{
\code{KorpusDirector} Class that directs the building of Korpus objects
}
\details{
The Document family of classes is an implementation of the builder
pattern documented in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). This pattern allows the
construction process to be defined at runtime.
}
\section{Korpus Director Methods}{

 \describe{
  \item{\code{new(builder)}}{Creates an object of Korpus Class}
  }
}

\seealso{
Other Corpus build family: \code{\link{KorpusBuilder0}},
  \code{\link{KorpusBuilder}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
