% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab.R
\docType{class}
\name{Lab}
\alias{Lab}
\title{Lab}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Lab
}
\arguments{
\item{name}{A character string containing the name of the Lab object. This variable is used in the instantiation and remove methods.}

\item{desc}{A chararacter string containing the description of the Lab}

\item{korpus}{An object of the Korpus class}

\item{visitor}{An object of one of the visitor classes.}
}
\description{
\code{Lab} Class that contains document corpora and the environment in which NLP happens.
}
\details{
The environment in which NLP happens. There are two groups of methods. The
first group allows clients to instantiate, retrieve, print, enter, leave,
and archive a Lab object.  The second set of methods allow clients to retrieve
the contained documents, add a document, and remove a document.

\strong{Lab Core Methods:}
 \describe{
  \item{\code{new(name, desc = NULL)}}{Creates an object of Lab Class}
  \item{\code{desc}}{A getter/setter method allowing clients to retrieve and set the Lab description variable.}
  \item{\code{getName()}}{Returns the name of the Lab object.}
  \item{\code{getPath()}}{Returns the path of the Lab object.}
  \item{\code{getLogs()}}{Returns the LogR object for the Lab object.}
  \item{\code{logIt(level = 'Info', fieldName = NA)}}{Formats the log and calls the LogR class to log an event.}
 }

\strong{Lab Aggregate Methods:}
 \describe{
  \item{\code{getKorpora()}}{Retrieves the list of corpora for the lab.}
  \item{\code{addKorpus(document)}}{Adds a corpus to the Lab object.}
  \item{\code{removeKorpus(document)}}{Removes a corpus from the Lab object. The corpus is archived in the NLPStudio archives.}
}

\strong{Lab Visitor Methods:}
 \describe{
  \item{\code{accept(visitor)}}{Accepts an object of the Visitor family of classes.}
}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
