% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logger.R
\docType{class}
\name{Logger}
\alias{Logger}
\title{Logger}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Logger
}
\description{
\code{Logger} Writes to log
}
\details{
Writes to log
}
\section{Class methods}{


\strong{Logger Core Methods:}
 \itemize{
  \item{\code{new()}}{Initiates thelogobject. }
  \item{\code{writeLog()}}{Writes log.}
  \item{\code{queryLog(...)}}{Enables client to perform queries on the log.}
}
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
