% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogR.R
\docType{class}
\name{LogR}
\alias{LogR}
\title{LogR}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LogR
}
\description{
\code{LogR} Writes to log
}
\details{
Writes to log
}
\section{Class methods}{


\strong{LogR Core Methods:}
 \itemize{
  \item{\code{new()}}{Initiates thelogobject. }
  \item{\code{writeLog()}}{Writes log.}
  \item{\code{queryLog(...)}}{Enables client to perform queries on the log.}
}
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
