% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\docType{class}
\name{Model}
\alias{Model}
\title{Model}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Model
}
\arguments{
\item{name}{A character string containing the name of the Model object. This variable is used in the instantiation and remove methods.}

\item{desc}{A chararacter string containing the description of the Model}

\item{corpus}{An object of the Corpus class}

\item{visitor}{An object of one of the visitor classes.}
}
\description{
\code{Model} Class in which corpora are created, processed, and transformed.
}
\details{
There are two groups of methods. The core methods allow clients to
instantiate models and to obtain their basic information.  Aggregate methods
enable clients to add, retrieve and remove corpora from the model.

\strong{Model Core Methods:}
 \describe{
  \item{\code{new(name, desc = NULL)}}{Creates an object of Model Class}
  \item{\code{desc}}{A getter/setter method allowing clients to retrieve and set the Model description variable.}
  \item{\code{getName()}}{Returns the name of the Model object.}
  \item{\code{getPath()}}{Returns the path of the Model object.}
  \item{\code{getLogs()}}{Returns the LogR object for the Model object.}
  \item{\code{logIt(level = 'Info', fieldName = NA)}}{Formats the log and calls the LogR class to log an event.}
 }

\strong{Model Aggregate Methods:}
 \describe{
  \item{\code{getCorpora()}}{Retrieves the list of corpora for the model.}
  \item{\code{addCorpus(corpus)}}{Adds a model to the Model object.}
  \item{\code{removeCorpus(corpus)}}{Removes a model from the Model object. The model is archived in the NLPStudios archives.}
}

\strong{Model Visitor Methods:}
 \describe{
  \item{\code{accept(visitor)}}{Accepts an object of the Visitor family of classes.}
}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
