% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingletonTemplate.R
\docType{class}
\name{NewSingletonClass}
\alias{NewSingletonClass}
\title{NewSingletonClass}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NewSingletonClass
}
\description{
\code{NewSingletonClass} Description goes here
}
\details{
More descriptive stuff

\strong{NewSingletonClass Core Methods:}
\describe{
 \item{\code{new()}}{Initializes the NewSingletonClass as s singleton object at load time.
}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
