% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLPStudio.R
\docType{class}
\name{NLPStudio}
\alias{NLPStudio}
\title{NLPStudio}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NLPStudio
}
\description{
\code{NLPStudio} Class the creates and manages labs.
}
\details{
This class creates and manages data labs  A data lab is essentially a
directory in which project data reside.  Multiple data labs can be created
to house separate versions of the data for analysis. Note: This class is a
singleton pattern. An NLPStudio object called nlpStudio is instantiated at
load time.  Any subsequent initializations will return the single nlpStudio
instance. There are two sets of methods.  The first set enables clients to
retrieve information about the NLPStudio object.  The second set allows
clients to add, remove, enter, and leave labs.

\strong{NLPStudio Core Methods:}
\describe{
 \item{\code{new()}}{Initializes the NLPStudio as s singleton object at load time.}
 \item{\code{getInstance()}}{Returns the current NLPStudio instance object. This will be the only instantiation called "nlpStudio.},
}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
