% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLPStudios.R
\docType{class}
\name{NLPStudios}
\alias{NLPStudios}
\title{NLPStudios}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NLPStudios
}
\description{
\code{NLPStudios} Class the creates and manages studios.
}
\details{
This class creates and manages data studios  A data studio is essentially a
directory in which project data reside.  Multiple data studios can be created
to house separate versions of the data for analysis. Note: This class is a
singleton pattern. An NLPStudios object called nlpStudios is instantiated at
load time.  Any subsequent initializations will return the single nlpStudios
instance. There are two sets of methods.  The first set enables clients to
retrieve information about the NLPStudios object.  The second set allows
clients to add, remove, enter, and leave studios.

\strong{NLPStudios Core Methods:}
\describe{
 \item{\code{new()}}{Initializes the NLPStudios as s singleton object at load time.}

\item{\code{getInstance()}}{Returns the current NLPStudios instance object. This will be the only instantiation called "nlpStudios.},
}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
