% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLPSuper.R
\docType{class}
\name{NLPSuper}
\alias{NLPSuper}
\title{NLPSuper}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NLPSuper
}
\description{
\code{NLPSuper} This is a super class that contains all common methods to
be implemented by all classes.
}
\details{
This class defines the common methods, such as logging and error handling
required by all classes
}
\section{Class methods}{


\strong{NLPSuper Core Methods:}
 \itemize{
  \item{\code{new(name, desc)}}{Instantiates class. Not implemented for this class.}
  \item{\code{logger(level, msg, fieldName)}}{Formats a log entry and passes it to the Logger object. }
}
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
