% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseSymbolCmd.R
\docType{class}
\name{ParseSymbolCmd}
\alias{ParseSymbolCmd}
\title{ParseSymbolCmd}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ParseSymbolCmd
}
\description{
\code{ParseSymbolCmd} Concrete command class responsible for executing the email parse command.
}
\details{
The Parser family of classes is an implementation of the command
pattern documented in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). This pattern allows one to
construct general components that need to delegate, sequence or execute
method calls at run time without the need to know the
class of the method or the method parameters.
}
\section{Parser Family Participants}{

 \itemize{
  \item{Command0}{Abstract class that defines the Commmand interface.}
  \item{Parser}{Invoker class that initiates the execution of a parse command. }
  \item{Document}{Receiver class with the parsing methods invoked by the Parser.}
  \item{ParseEmailCmd}{Concrete Command0 class for parsing email addresses from a Document object.}
  \item{ParseUrlCmd}{Concrete Command0 class for parsing URLs addresses from a Document object.}
  \item{ParseTwitterCmd}{Concrete Command0 class for parsing twitter handles  from a Document object.}
  \item{ParseControlCmd}{Concrete Command0 class for parsing control characters  from a Document object.}
  \item{ParseHyphenCmd}{Concrete Command0 class for parsing hyphens  from a Document object.}
  \item{ParseAposCmd}{Concrete Command0 class for parsing apostrophes  from a Document object.}
  \item{ParsePunctCmd}{Concrete Command0 class for parsing punctuation  from a Document object.}
  \item{ParseSymbolCmd}{Concrete Command0 class for parsing symbols  from a Document object.}
  \item{ParseDigitCmd}{Concrete Command0 class for parsing digits  from a Document object.}
  \item{ParseRepeatCmd}{Concrete Command0 class for parsing repeated chars  from a Document object.}
  \item{ParseLongWordCmd}{Concrete Command0 class for parsing long words  from a Document object.}#'
 }
}

\section{ParseSymbolCmd methods}{

 \itemize{
  \item{\code{initialize(set)}}{Accepts a Set Object containing Document receiver objects.}
  \item{\code{execute()}}{Iterates through a Set objects Document objects and executes the request.}
 }
}

\seealso{
Other Parser classes: \code{\link{ParseControlCmd}},
  \code{\link{ParseDigitCmd}}, \code{\link{ParseEmailCmd}},
  \code{\link{ParseHyphenCmd}},
  \code{\link{ParseLongWordCmd}},
  \code{\link{ParsePunctCmd}},
  \code{\link{ParseRepeatCmd}},
  \code{\link{ParseTwitterCmd}}, \code{\link{ParseUrlCmd}},
  \code{\link{Parser}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
