% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipeline.R
\docType{class}
\name{Pipeline}
\alias{Pipeline}
\title{Pipeline}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Pipeline
}
\arguments{
\item{name}{A character string containing the name of the Pipeline object. This variable is used in the instantiation and remove methods.}

\item{desc}{A chararacter string containing the description of the Pipeline}

\item{visitor}{An object of one of the visitor classes.}
}
\description{
\code{Pipeline} Class containing the NLP processing, analysis, and modeling pipeline.
}
\details{
The pipeline starts with the extraction of data from external sources and includes
the establishment of the raw data sets, the cross-validation sets, data
preprocessing, analysis, feature engineering and selection, modeling
an model evaluation.

\strong{Pipeline Methods:}
 \describe{
  \item{\code{new(name, desc = NULL)}}{Creates an object of Pipeline Class}
  \item{\code{desc}}{A getter/setter method allowing clients to retrieve and set the Pipeline description variable.}
  \item{\code{getName()}}{Returns the name of the Pipeline object.}
  \item{\code{getPath()}}{Returns the path of the Pipeline object.}
  \item{\code{accept(visitor)}}{Accepts an object of the Visitor family of classes.}
  \item{\code{logIt(level = 'Info', fieldName = NA)}}{Formats the log and calls the LogR class to log an event.}
}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
