% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\docType{class}
\name{Preprocess}
\alias{Preprocess}
\title{Preprocess}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Preprocess
}
\arguments{
\item{document}{Object of the Document class.}

\item{malContent}{Data frame of words to be extracted from the text}

\item{normalizations}{Data frame of key value pairs of patterns and replacements}
}
\description{
\code{Preprocess} Class that performs document preprocessing tasks.
}
\details{
Class responsible for document preprocessing tasks,such as encoding,
normalization, sanitization, and tokenization of text.
}
\section{Document methods}{

 \itemize{
  \item{\code{new()}}{Instantiates an object of the Preprocess class .}
  \item{\code{encode()}}{Method for repairing and converting document encoding.}
  \item{\code{normalize()}}{Method for normalizing text,e.g. abbrevaitions, common misspellings.}
  \item{\code{extract()}}{Method for extracting words or sentences including select words from the text.}
 }
}

\seealso{
Other Document classes: \code{\link{Document}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
