% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read0.R
\docType{class}
\name{Read0}
\alias{Read0}
\title{Read0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Read0
}
\arguments{
\item{path}{Character string indicating the path to the document to be read}

\item{header}{Logical indicating whether the csv file contains headers.  Used for the ReadCsv class.}
}
\description{
\code{Read0} Class as in abstract class to the ReadBin, ReadText, ReadCsv, ReadRdata read strategies
}
\details{
This class is an abstract class defining the methods to be implemented by
the concrete strategy classes: ReadText, ReadBin, ReadCsv, and ReadRdata.
}
\section{Methods}{

\describe{
 \item{\code{new()}}{Returns an object of the selected concrete read strategy class. This method is overwritten by the concrete strategy classes}
 \item{\code{readData(path)}}{Returns the content of the file with its location designated in the path parameter.}
}
}

\author{
John James, \email{j2sdatalab@gmail.com}
}
\keyword{datasets}
