% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadBin.R
\docType{class}
\name{ReadBin}
\alias{ReadBin}
\title{ReadBin}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ReadBin
}
\arguments{
\item{path}{Character string containing the path to the document to be read.}
}
\description{
\code{ReadBin} Class contains the data and methods for reading files in binary format.
}
\details{
This class inherits from the Read0 inteface class and provides the data
and methods for reading files in binary format.
}
\section{Methods}{

\describe{
 \item{\code{readData(path)}}{Reads data from the location designated in the path parameter}
}
}

\seealso{
Other Document i/o classes: \code{\link{WriteBin}}
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
\keyword{datasets}
