% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadCsv.R
\docType{class}
\name{ReadCsv}
\alias{ReadCsv}
\title{ReadCsv}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ReadCsv
}
\arguments{
\item{path}{A character string indicating the location for the file to be read.}

\item{header}{A logical indicating whether the csv file contains headers. Defaults to TRUE}
}
\description{
\code{ReadCsv} Class contains the data and methods for reading csv files
}
\details{
This class inherits from the Read0 inteface class and provides the data
and methods for reading csv files
}
\section{Methods}{

\describe{
 \item{\code{readData(directory, fileName)}}{Reads data from the designated  directory and file name}
}
}

\seealso{
Other Document i/o classes: \code{\link{ReadBin}},
  \code{\link{ReadRdata}}, \code{\link{ReadText}},
  \code{\link{Write0}}, \code{\link{WriteBin}},
  \code{\link{WriteCsv}}, \code{\link{WriteRdata}},
  \code{\link{WriteText}}
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
\keyword{datasets}
