% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadRdata.R
\docType{class}
\name{ReadRdata}
\alias{ReadRdata}
\title{ReadRdata}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ReadRdata
}
\arguments{
\item{path}{Character string containing the path to the document to be read.}
}
\description{
\code{ReadRdata} Class contains the data and methods for reading Rdata files
}
\details{
This class inherits from the Read0 inteface class and provides the data
and methods for reading Rdata data from files
}
\section{Fields}{

\describe{
\item{\code{objectName}}{A character string containing the name to assign to the Rdata object}
}}

\section{Fields}{

}

\section{Methods}{

\describe{
 \item{\code{readData(directory, fileName)}}{Reads Rdata file from the designated  directory and file name and assigns the object name in the global environment }
}
}

\seealso{
Other Document i/o classes: \code{\link{ReadBin}},
  \code{\link{ReadCsv}}, \code{\link{ReadText}},
  \code{\link{Write0}}, \code{\link{WriteBin}},
  \code{\link{WriteCsv}}, \code{\link{WriteRdata}},
  \code{\link{WriteText}}
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
\keyword{datasets}
