% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadText.R
\docType{class}
\name{ReadText}
\alias{ReadText}
\title{ReadText}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ReadText
}
\arguments{
\item{path}{Character string with location of the file to be read.}
}
\description{
\code{ReadText} Class contains the data and methods for reading text
}
\details{
This class inherits from the Read0 inteface class and provides the data
and methods for reading text.
}
\section{Methods}{

\describe{
 \item{\code{readData(directory, fileName)}}{Reads text from the designated  directory and file name}
}
}

\seealso{
Other Document i/o classes: \code{\link{ReadBin}},
  \code{\link{ReadCsv}}, \code{\link{ReadRdata}},
  \code{\link{Write0}}, \code{\link{WriteBin}},
  \code{\link{WriteCsv}}, \code{\link{WriteRdata}},
  \code{\link{WriteText}}
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
\keyword{datasets}
