% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Register.R
\docType{class}
\name{Register}
\alias{Register}
\title{Register}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Register
}
\description{
\code{Register} Class containing methods for interacting with corpus registers.
}
\details{
This leaf class of the Document0 composite pattern provides methods for
obtaining and printing the register meta data as well as reading and writing
the document.
}
\section{Core Methods}{

The core methods are shared by both primative and composite objects and are
as follows:
\describe{
 \item{getDocument}{Base method for retrieving the document meta data. Specific behaviors implemented in the Register and Korpus subclasses }
 \item{printDocument}{Base method for printing the document meta data to the console. Specific behaviors implemented in the Register and Korpus subclasses}
 \item{readDocument}{Base method for reading documents. Specific behaviors implemented in the Register and Korpus subclasses}
 \item{writeDocument}{Base method for writing documents. Specific behaviors implemented in the Register and Korpus subclasses}
 \item{archiveDocument}{Base method for archiving documents. Specific behaviors implemented in the Register and Korpus subclasses}
}
}

\section{Behavior Methods}{

These methods enable clients to define behaviors for Register and Korpus class object at runtime. These methods are:
\describe{
 \item{getReader}{Base method for returning the current Reader subclass for the object.}
 \item{setReader}{Base method for setting the current Reader subclass for the object.}
 \item{getWriter}{Base method for returning the current Writer subclass for the object.}
 \item{setWriter}{Base method for setting the current Writer subclass for the object.}
}
}

\author{
John James, \email{j2sdatalab@gmail.com}
}
\keyword{datasets}
