% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set.R
\docType{class}
\name{Set}
\alias{Set}
\title{Set}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Set
}
\arguments{
\item{name}{A character string containing the name of the Set object. This variable is used in the instantiation and remove methods.}

\item{desc}{A chararacter string containing the description of the Set.}

\item{document}{An object of one of the Document sub-classes.}

\item{visitor}{An object of one of the visitor classes.}
}
\description{
\code{Set} Class that defines a set or collection of documents
}
\details{
Class contains a collection of documents of various types, including text
documents, nGrams, and pos tagged documents.
}
\section{Set Core Methods}{

 \describe{
  \item{\code{new(name, desc = NULL, lab = NULL)}}{Creates an object of Set Class}
  \item{\code{desc}}{A getter/setter method allowing clients to retrieve and set the Set description variable.}
  \item{\code{parent}}{A getter/setter method allowing clients to retrieve and set the Lab object to which the Set object belongs.}
  \item{\code{getName()}}{Returns the name of the Set object.}
  \item{\code{getPath()}}{Returns the path of the Set object.}
  \item{\code{getDocuments()}}{Returns the list of Document class objects.}
  \item{\code{addDocument(document)}}{Adds a Document object to the set.}
  \item{\code{removeDocument(document)}}{Removes a Document object from the set.}
  \item{\code{logIt(level = 'Info', fieldName = NA)}}{Formats the log and calls the LogR class to log an event.}
  \item{\code{accept(visitor)}}{Accepts an object of the Visitor family of classes.}
 }
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
