% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Singleton.R
\docType{class}
\name{Singleton}
\alias{Singleton}
\title{Singleton}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Singleton
}
\description{
\code{Singleton} Singleton class for NLPStudios subclass
}
\details{
This class ...
}
\section{Methods}{

\describe{
 \item{\code{initialize(...)}}{Initializes a Singleton object}
 \item{\code{getInstance(...)}}{Gets instance of singleton object }
}
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
