% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snap0.R
\docType{class}
\name{Snap0}
\alias{Snap0}
\title{Snap0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Snap0
}
\arguments{
\item{object}{The object for which the snapshot is taken and saved to disk.}

\item{snapshotName}{The name of a snapshot to be restored from disk.}
}
\description{
\code{Snap0} Is an abstract class that manages the process of saving and
restoring snapshots of NLPStudio objects.
}
\details{
\strong{Snap0 Class Overview:}

The Snap0 singleton class defines the interface for saving and restoring snapshots
of NLPStudio objects. Its single object delegates the actual save and restore
operations to the SnapSave and SnapRestore classes, respectively. Once the
request has been fulfilled by the subclasses, the request is logged.

\strong{Snap0 Class Collaborators:}
The collaborators of the Snap0 class are:
\itemize{
 \item Lab: Class in which NLP takes place. The class contains objects of the DocumentCollection class.
 \item DocumentCollection: Composite class containing objects of the Document class.
 \item Document: Leaf class of the Document objects.
 }

\strong{Snap Family of Classes Participants:}
\itemize{
 \item Snap0: This class which manages and reports the saving and restoration of snapshots.
 \item SnapSave: Class which saves the snapshot of files associated with Lab objects to the lab snapshots directory disk.
 \item SnapRestore: Class which restores snapshots associated with Lab objects from disk.
 }

\strong{Methods:}
\describe{
 \item{\code{new()}}{Instantiates the Snap0 object. This singleton object is instantiated by the NLPStudio object at package load time.}
 \item{\code{save(object)}}{Executes the save operation using the current saver method.}
 \item{\code{restore(snapshotName)}}{Executes the restore operation using the current restorer method.}
 \item{\code{accept(self)}}{Accepts the dispatch from a visitor and dispatches the appropriate visitor method, sending itself as a parameter.}
 \item{\code{getSnapshots(object)}}{Returns a list of snapshots for a specific object.}
}
}
\seealso{
Other Snapshot classes: \code{\link{SnapSave}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
