% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnapSave.R
\docType{class}
\name{SnapSave}
\alias{SnapSave}
\title{SnapSave}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SnapSave
}
\arguments{
\item{snapName}{Character string indicating the unique name for the snapshot.}

\item{object}{Object of the "Corpus", "Document", "DocumentCollection", or "Lab" class to be archived.}
}
\description{
\code{SnapSave} Class for saving snapshots of objects
}
\details{
\strong{SnapSave Class Overview:}

The SnapSave class manages the process of archiving and restoring objects
in the NLPStudio.

\strong{Snap Family of Classes Participants:}
\itemize{
 \item Snap0: This class which manages and reports the saving and restoration of snapshots.
 \item SnapSave: Class which saves the snapshot to disk.
 \item SnapRestore: Class which restores a snapshot from disk.
 }

\strong{SnapSave Methods:}
\itemize{
 \item{\code{save(object)}}{Method that saves the snapshot of an object. A directory is created containing two compressed files: (1), the serialized object, and (2), the files in the object's directory.}
}
}
\seealso{
Other Snapshot classes: \code{\link{Snap0}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
