% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source0.R
\docType{class}
\name{Source0}
\alias{Source0}
\title{Source0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Source0
}
\description{
\code{Source0}{Abstract class for obtaining data from varied sources}
}
\details{
Abstract class that defines the data and methods used to
obtain text data from varied sources, such as the internet or other
collections within the studio. Future extensions will provide methods
for xml parsing and more sophisticated web scraping.
}
\section{Methods}{

\describe{
 \item{\code{new()}}{Instantiates an object of type Source0. This method is overwritten by concrete strategy classes.}
 \item{\code{getData(...)}}{Method that obtains the data by executing the source strategy}
}
}

\author{
John James, \email{j2sdatalab@gmail.com}
}
\keyword{datasets}
