% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceData0.R
\docType{class}
\name{SourceData0}
\alias{SourceData0}
\title{SourceData0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SourceData0
}
\value{
Character string indicating the name of the Validator class.
}
\description{
\code{SourceData0} Abstract class for the SourceData family of classes
}
\details{
Abstract class for the SourceData classes
}
\section{Methods}{

\describe{
 \item{\code{initialize()}}{Not implemented for this abstract class.}
 \item{\code{logIt()}}{Not implemented for this abstract class.}
}
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
