% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceDataWeb.R
\docType{class}
\name{SourceDataWeb}
\alias{SourceDataWeb}
\title{SourceDataWeb}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SourceDataWeb
}
\arguments{
\item{name}{Character string indicating the name of the SourceDataWeb object.}

\item{desc}{Character string containing the description of the corpus.}

\item{pipeline}{Pipeline object into to which the data is to be used.}

\item{url}{Character string indicating the url from which the data will be sourced.}
}
\description{
\code{SourceDataWeb} Class responsible for obtaining corpus data from web sources.
}
\details{
The corpus data are downloaded into the data/external/name subdirectory for
the designated pipeline and name.

The SourceDataWeb family of classes is an implementation of the strategy
pattern documented in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). This pattern allows corpus
sourcing behavior to be defined at run time.
}
\section{SourceDataWeb Family Participants}{

 \itemize{
  \item{SourceData0}{Abstract class that defines core methods.}
  \item{SourceDataWeb}{Class responsible for obtaining corpus data from web sources.}
 }
}

\section{SourceDataWeb methods}{

 \itemize{
  \item{\code{new(...)}}{Method not implemented for this abstract class.}
  \item{\code{sourceData()}}{Method not implemented for this abstract class.}
  \item{\code{getURL()}}{Method used to return the URL, primarily for validation purposes. }
  \item{\code{accept()}}{Method used to accept visitor objects. }
 }
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
