% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceWeb.R
\docType{class}
\name{SourceWeb}
\alias{SourceWeb}
\title{SourceWeb}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SourceWeb
}
\arguments{
\item{downloadParams}{A list containing the following fields:
\itemize{
 \item{url}{The URL from which the content will be downloaded.}
 \item{destFile}{The file path into which the data will be downloaded.}
}}

\item{unZipParams}{A list containing the following fields:
\itemize{
 \item{zipFiles}{A character vector including the files to be extracted from the zip file.}#'
 \item{unZipPath}{A character string containing the directory into which the zip files are to be extracted.}
 \item{zipPath}{A character string containing the file path to the zipped file.}
}}
}
\description{
\code{SourceWeb}{Class with methods for obtaining data from internet sources}
}
\details{
Concrete strategy class for obtaining data from web source.  This includes
methods for downloading and unzipping data.
}
\section{Methods}{

\describe{
 \item{\code{new()}}{Instantiates an object of type SourceWeb.}
 \item{\code{getData(...)}}{Method that downloads and unzips the data}
}
}

\section{Parameters}{

}

\author{
John James, \email{j2sdatalab@gmail.com}
}
\keyword{datasets}
