% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/State.R
\docType{class}
\name{State}
\alias{State}
\title{State}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
State
}
\arguments{
\item{dateFrom}{A parameter of the query method. An ISO 8601 formatted date indicating the date from which states should be returned.}

\item{dateTo}{A parameter of the query method. An ISO 8601 formatted date indicating the date to which states should be returned.}

\item{class}{A parameter of the query method. A character string indicating the class of state objects that should be returned.}

\item{objectName}{A parameter of the query method. A character string indicating the object for which states should be returned.}

\item{object}{A serializable object to be serialized or deserialized.}
}
\description{
\code{State} This class takes receives query, save, and restore requests
from client applications and visitor methods, validates the request, and
dispatches the appropriate visitor class for fullfillment.
}
\details{
The State class is a participant of the State family of classes.

\strong{State Family of Classes Overview:}
The State family of classes is responsible for object persistence and
data recovery in the NLPStudio package and is comprised of the following
participants.

\strong{State Class Family Participants:}
\itemize{
 \item State: This class oversees the process of saving, restoring and
 querying states. It takes requests from client applications and dispatches
 instantiates the appropriate visitor to fulfill the request. This class
 also provides state query functionality.
 \item Curator: This class oversees the process of restoring a composite
 to a prior state.
 \item StateManager: This class has a persistent single object that
 receives a save/restore request from a visitor, keeps track of the
 saved states, and their locations, and dispatches the appropriate
 save/restore method within the State class to serialize or
 deserialize the object.
 \item StateServer: This class is responsible for serializing and deserializing
 object states.
 }

\strong{State Class Methods:}
The following methods are defined for this class:
\describe{
\item{\code{new()}}{Method for instantiating objects of the State class. Obtains the list of serializeable classes for validation puroses.}
 \item{\code{query(dateFrom, dateTo, class, objectName)}}{Method for quering the saved states. Returns a list of states matching the query parameters.}
 \item{\code{save(object)}}{Method for saving the state of a serializable object. This method invokes the accept method on the object which dispatches the appropriate save visitor.}
 \item{\code{restore(object)}}{Method for restoring an the designated object to the state identified by the stateId parameter. The method invokes the accept method on the object, which dispatches the appropriate restore visitor to obtain the prior state, then invokes the Curator to return the composite to the prior state}
 }
}
\seealso{
Other State Classes: \code{\link{StateManager}},
  \code{\link{StateServer}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
