% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudioCache.R
\docType{class}
\name{StudioCache}
\alias{StudioCache}
\title{StudioCache}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
StudioCache
}
\description{
\code{Cache} Class for managing the cache for all NLP objects.
}
\details{
Class and methods to create a cache environment, retrieve objects from
the cache, and write objects to the cache.  There are also methods for
obtaining cached objects from a file and generating a list of the contents
of the cache environment.
}
\section{Methods}{

\describe{
 \item{\code{initialize()}}{Creates the cache environment.}
 \item{\code{getCache(key)}}{Retrieves object from cache.}
 \item{\code{setCache(key, value)}}{Save a key value pair to the cache environment.}
 \item{\code{purgeCache()}}{Purges the entire cache.}
 \item{\code{loadCache()}}{Loads the cache from file.}
 \item{\code{saveCache()}}{Restores items in cache to the global environment.}
 \item{\code{restoreCache()}}{Purges the entire cache.}
 \item{\code{lsCache}}{Lists the names of the items in the cache environment.}
 \item{\code{printCache()}}{Prints the contents of the cache environment.}
 \item{\code{getAllCache()}}{Returns the entire cache environment.}
 \item{\code{getInstance()}}{Returns the instance of the singleton StudioCache class.}
}
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
