% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudioCacheManager.R
\docType{class}
\name{StudioCacheManager}
\alias{StudioCacheManager}
\title{StudioCacheManager}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
StudioCacheManager
}
\description{
\code{Cache} Class for interacts with the StudioCache class to ensure that the cache is saved to disk whenever the cache has changed
}
\details{
Class and methods to read, write, and save objects to, from and of the StudioCache.
}
\section{Methods}{

\describe{
 \item{\code{initialize()}}{Creates the cache environment.}
 \item{\code{getCache(key)}}{Retrieves object from cache.}
 \item{\code{setCache(key, value)}}{Save a key value pair to the cache environment.}
 \item{\code{getInstance()}}{Returns the instance of the singleton StudioCache class.}
}
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
