% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\docType{class}
\name{Tools}
\alias{Tools}
\title{Tools}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Tools
}
\arguments{
\item{len}{The length of the requested random string.}
}
\description{
\code{Tools} Class containing various helper methods used within the package.
}
\section{Tools Methods}{

\describe{
 \item{\code{makeRandomString(len = 12)}}{Returns a random string of the requested length.}
}
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
