% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateClass.R
\docType{class}
\name{ValidateClass}
\alias{ValidateClass}
\title{ValidateClass}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidateClass
}
\value{
A logical TRUE if class is equal to expected value, FALSE otherwise.
}
\description{
\code{ValidateClass} Class for validating object class
}
\details{
This class provide a methods for validating an object's class.
}
\section{Methods}{

\describe{
 \item{\code{new()}}{Creates an object of ValidateClass class}
 \item{\code{validate(object)}}{Method for validating logical}
}
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
