% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateExists.R
\docType{class}
\name{ValidateExists}
\alias{ValidateExists}
\title{ValidateExists}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidateExists
}
\value{
A logical  TRUE if the object exists, FALSE otherwise
}
\description{
\code{ValidateExists} Class for validating whether an object exists
}
\details{
This class provide a methods for validating the existence of objects in
the global environment
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
