% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateLogical.R
\docType{class}
\name{ValidateLogical}
\alias{ValidateLogical}
\title{ValidateLogical}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidateLogical
}
\value{
A logical TRUE if valid logical, FALSE otherwise
}
\description{
\code{ValidateLogical} Class for validating logicals
}
\details{
This class provide a methods for validating logicals.
}
\section{Methods}{

\describe{
 \item{\code{new()}}{Creates an object of ValidateLogical class}
 \item{\code{validate(object)}}{Method for validating logical}
}
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
