% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateName.R
\docType{class}
\name{ValidateName}
\alias{ValidateName}
\title{ValidateName}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidateName
}
\arguments{
\item{cls}{Character string containing the class invoking the validation.}

\item{method}{Character string containing the name of the invoking method}

\item{name}{Character string containing the name variable}

\item{expect}{Logical c(TRUE, FALSE) indicating expected outcome}
}
\value{
Throws an error if the field doesn't pass validation
}
\description{
\code{ValidateName} This class validates names of objects
}
\section{Public Methods}{

\describe{
 \item{\code{new()}}{Creates an object of ValidateName class}
 \item{\code{validate(cls, method, name, expect)}}{Validates the name variable}
}
}

\section{Parameters}{

}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
