% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateNoSpaces.R
\docType{class}
\name{ValidateNoSpaces}
\alias{ValidateNoSpaces}
\title{ValidateNoSpaces}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidateNoSpaces
}
\value{
a logical TRUE if valid string, FALSE otherwise
}
\description{
\code{ValidateNoSpaces} Class for validating string has no spaces
}
\details{
This class provide a methods for validating character string has no spaces
}
\section{Methods}{

\describe{
 \item{\code{new()}}{Creates an object of ValidateNoSpaces class}
 \item{\code{validate(object)}}{Method for validating string}
}
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
