% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateNotEmpty.R
\docType{class}
\name{ValidateNotEmpty}
\alias{ValidateNotEmpty}
\title{ValidateNotEmpty}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidateNotEmpty
}
\value{
a logical TRUE if not empty, FALSE otherwise
}
\description{
\code{ValidateNotEmpty} Class for validating string is not empty
}
\details{
This class provide a methods for validating character string is not empty
}
\section{Methods}{

\describe{
 \item{\code{new()}}{Creates an object of ValidateNotEmpty class}
 \item{\code{validate(object)}}{Method for validating string}
}
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
