% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateNumeric.R
\docType{class}
\name{ValidateNumeric}
\alias{ValidateNumeric}
\title{ValidateNumeric}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidateNumeric
}
\value{
A logical TRUE if valid numeric, FALSE otherwise
}
\description{
\code{ValidateNumeric} Class for validating numerics
}
\details{
This class provide a methods for validating numerics.
}
\section{Methods}{

\describe{
 \item{\code{new()}}{Creates an object of ValidateNumeric class}
 \item{\code{validate(object)}}{Method for validating numeric}
}
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
