% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidatePath.R
\docType{class}
\name{ValidatePath}
\alias{ValidatePath}
\title{ValidatePath}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidatePath
}
\arguments{
\item{cls}{Character string containing the name of the invoking class}

\item{method}{Character string containing the name of the invoking method}

\item{fieldName}{Character string containing the name of the field to validate}

\item{value}{The value of the field being validated}

\item{level}{Character string c("Error", "Warn", "Info)}

\item{msg}{Character vector containing the message to be displayed if test fails}

\item{expect}{Logical c(TRUE, FALSE) indicating expected outcome}
}
\value{
A logical, TRUE if outcome was expected, FALSE otherwise
}
\description{
\code{ValidatePath} Class for validating a file path

This class provide a methods for validating file paths.
}
\section{Private Methods}{

\describe{
 \item{\code{notifyInfo(note)}}{Renders a notification with level "info"}
 \item{\code{notifyWarn(note)}}{Renders a notification with level "warn"}
 \item{\code{notifyError(note)}}{Renders a notification with level "error", and stops processing}
}
}

\section{Public Methods}{

\describe{
 \item{\code{new()}}{Creates an object of Validate class}
 \item{\code{validate(cls, method, fieldName, value, level, msg, expect)}}{Validation method overwritten by subclasses}
 \item{\code{notify(cls, method, fieldName, value, level, msg, expect)}}{Renders a notification if outcome is not expected}
}
}

\section{Parameters}{

}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
