% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateString.R
\docType{class}
\name{ValidateString}
\alias{ValidateString}
\title{ValidateString}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidateString
}
\value{
a logical TRUE if valid string, FALSE otherwise
}
\description{
\code{ValidateString} Class for validating strings
}
\details{
This class provide a methods for validating character strings
}
\section{Methods}{

\describe{
 \item{\code{new()}}{Creates an object of ValidateString class}
 \item{\code{validate(object)}}{Method for validating string}
}
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
