% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateUrl.R
\docType{class}
\name{ValidateUrl}
\alias{ValidateUrl}
\title{ValidateUrl}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidateUrl
}
\value{
A logical TRUE if valid logical, FALSE otherwise
}
\description{
\code{ValidateUrl} Class for validating website urls
}
\details{
This class provide a methods for validating logicals.
}
\section{Methods}{

\describe{
 \item{\code{new()}}{Creates an object of ValidateUrl class}
 \item{\code{validate(object)}}{Method for validating logical}
}
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
