% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validator0.R
\docType{class}
\name{Validator0}
\alias{Validator0}
\title{Validator0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Validator0
}
\value{
Character string indicating the name of the Validator class.
}
\description{
\code{Validator0} Abstract class for the Validator classes
}
\details{
Abstract class for the Validator classes
}
\section{Methods}{

\describe{
 \item{\code{getName()}}{Method that returns the name of the validator.}
}
}

\seealso{
Other Validation Classes: \code{\link{VValidatorCorpusBuilder}},
  \code{\link{ValidatorClass}},
  \code{\link{ValidatorDate}},
  \code{\link{ValidatorExists}}, \code{\link{ValidatorIO}},
  \code{\link{ValidatorLogical}},
  \code{\link{ValidatorNoSpaces}},
  \code{\link{ValidatorNotEmpty}},
  \code{\link{ValidatorNumeric}},
  \code{\link{ValidatorPath}},
  \code{\link{ValidatorString}},
  \code{\link{ValidatorUrl}}, \code{\link{Validator}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
