% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidatorKorpusBuilder.R
\docType{class}
\name{ValidatorBuilder}
\alias{ValidatorBuilder}
\title{VValidatorBuilder}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidatorBuilder
}
\arguments{
\item{object}{The object in its current state}

\item{...}{Parameters}
}
\description{
\code{VValidatorBuilder} Visitor class responsible for validating the initialization objects of all classes
}
\details{
\strong{VValidatorBuilder Methods:}
The VValidatorBuilder methods are as follows:
 \itemize{
  \item{\code{nlpStudio(object)}}{Method for validating the instantiation of the NLPStudio object}
  \item{\code{lab(object)}}{Method for validating the instantiation of the Lab object}
  \item{\code{documentCollection(object)}}{Method for validating the instantiation of the DocumentCollection object.}
  \item{\code{documentText(object)}}{Method for validating the instantiation of the DocumentText object.}
  \item{\code{documentCsv(object)}}{Method for validating the instantiation of the DocumentCsv object.}
  \item{\code{documentRdata(object)}}{Method for validating the instantiation of the DocumentRdata object.}
  \item{\code{documentXlsx(object)}}{Method for validating the instantiation of the DocumentXlsx object.}
}
}
\seealso{
Other Validation Classes: \code{\link{Validator0}},
  \code{\link{ValidatorClass}},
  \code{\link{ValidatorDate}},
  \code{\link{ValidatorExists}},
  \code{\link{ValidatorLogical}},
  \code{\link{ValidatorNoSpaces}},
  \code{\link{ValidatorNotEmpty}},
  \code{\link{ValidatorNumeric}},
  \code{\link{ValidatorPath}},
  \code{\link{ValidatorString}},
  \code{\link{ValidatorUrl}}, \code{\link{Validator}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
