% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidatorIO.R
\docType{class}
\name{ValidatorIO}
\alias{ValidatorIO}
\title{ValidatorIO}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidatorIO
}
\arguments{
\item{value}{Character string containing the value to be validated}
}
\value{
A logical TRUE if class is equal to expected value, FALSE otherwise.
}
\description{
\code{ValidatorIO} Class for validating the io method for an object.
}
\details{
Class compares the file type with the IO Class assigned to the object and
returns FALSE if there is a mismatch.
}
\section{Parameters}{

}

\seealso{
Other Validation Classes: \code{\link{VValidatorCorpusBuilder}},
  \code{\link{Validator0}}, \code{\link{ValidatorClass}},
  \code{\link{ValidatorDate}},
  \code{\link{ValidatorExists}},
  \code{\link{ValidatorLogical}},
  \code{\link{ValidatorNoSpaces}},
  \code{\link{ValidatorNotEmpty}},
  \code{\link{ValidatorNumeric}},
  \code{\link{ValidatorPath}},
  \code{\link{ValidatorString}},
  \code{\link{ValidatorUrl}}, \code{\link{Validator}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
