% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidatorName.R
\docType{class}
\name{ValidatorName}
\alias{ValidatorName}
\title{ValidatorName}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidatorName
}
\arguments{
\item{class}{Character string containing the class invoking the validation.}

\item{method}{Character string containing the name of the invoking method}

\item{name}{Character string containing the name variable}

\item{expect}{Logical c(TRUE, FALSE) indicating expected outcome}
}
\value{
Throws an error if the field doesn't pass validation
}
\description{
\code{ValidatorName} This class validates names of objects
}
\section{Public Methods}{

\describe{
 \item{\code{new()}}{Creates an object of ValidatorName class}
 \item{\code{validate(class, method, name, expect)}}{Validates the name variable}
}
}

\section{Parameters}{

}

\seealso{
Other Validation Classes: \code{\link{Authorizer}},
  \code{\link{Validator0}}, \code{\link{ValidatorClass}},
  \code{\link{ValidatorDate}},
  \code{\link{ValidatorExists}},
  \code{\link{ValidatorLogical}},
  \code{\link{ValidatorNoSpaces}},
  \code{\link{ValidatorNotEmpty}},
  \code{\link{ValidatorNumeric}},
  \code{\link{ValidatorPath}},
  \code{\link{ValidatorString}},
  \code{\link{ValidatorUrl}}, \code{\link{Validator}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
