% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidatorNumeric.R
\docType{class}
\name{ValidatorNumeric}
\alias{ValidatorNumeric}
\title{ValidatorNumeric}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidatorNumeric
}
\arguments{
\item{value}{Character string containing the value to be validated}

\item{expect}{Character string or logical indicating the expected value}
}
\value{
A logical TRUE if class is equal to expected value, FALSE otherwise.
}
\description{
\code{ValidatorNumeric} Class for validating numerics
}
\details{
This class provide a methods for validating numerics.
}
\section{Methods}{

\describe{
 \item{\code{validate(value, expect = NULL)}}{Method for validating numeric}
}
}

\section{Parameters}{

}

\seealso{
Other Validation Classes: \code{\link{IOFactory}},
  \code{\link{Validator0}}, \code{\link{ValidatorClass}},
  \code{\link{ValidatorDate}},
  \code{\link{ValidatorExists}},
  \code{\link{ValidatorLogical}},
  \code{\link{ValidatorNoSpaces}},
  \code{\link{ValidatorNotEmpty}},
  \code{\link{ValidatorString}},
  \code{\link{ValidatorUrl}}, \code{\link{Validator}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
