% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidatorPath.R
\docType{class}
\name{ValidatorPath}
\alias{ValidatorPath}
\title{ValidatorPath}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidatorPath
}
\arguments{
\item{value}{Character string containing the value to be validated}

\item{expect}{Character string or logical indicating the expected value}
}
\value{
A logical TRUE if class is equal to expected value, FALSE otherwise.
}
\description{
\code{ValidatorPath} Class for validating a file path

This class provide a methods for validating file paths.
}
\section{Parameters}{

}

\seealso{
Other Validation Classes: \code{\link{Validator0}},
  \code{\link{ValidatorBuilder}},
  \code{\link{ValidatorClass}},
  \code{\link{ValidatorDate}},
  \code{\link{ValidatorExists}},
  \code{\link{ValidatorLogical}},
  \code{\link{ValidatorNoSpaces}},
  \code{\link{ValidatorNotEmpty}},
  \code{\link{ValidatorNumeric}},
  \code{\link{ValidatorString}},
  \code{\link{ValidatorUrl}}, \code{\link{Validator}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
