% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidatorString.R
\docType{class}
\name{ValidatorString}
\alias{ValidatorString}
\title{ValidatorString}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidatorString
}
\arguments{
\item{value}{Character string containing the value to be validated}

\item{expect}{Character string or logical indicating the expected value}
}
\value{
A logical TRUE if class is equal to expected value, FALSE otherwise.
}
\description{
\code{ValidatorString} Class for validating strings
}
\details{
This class provide a methods for validating character strings
}
\section{Methods}{

\describe{
 \item{\code{validate(value, expect = NULL)}}{Method for validating string}
}
}

\section{Parameters}{

}

\seealso{
Other Validation Classes: \code{\link{Validator0}},
  \code{\link{ValidatorBuilder}},
  \code{\link{ValidatorClass}},
  \code{\link{ValidatorDate}},
  \code{\link{ValidatorExists}},
  \code{\link{ValidatorLogical}},
  \code{\link{ValidatorNoSpaces}},
  \code{\link{ValidatorNotEmpty}},
  \code{\link{ValidatorNumeric}},
  \code{\link{ValidatorPath}}, \code{\link{ValidatorUrl}},
  \code{\link{Validator}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
