% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidatorUrl.R
\docType{class}
\name{ValidatorUrl}
\alias{ValidatorUrl}
\title{ValidatorUrl}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidatorUrl
}
\value{
A logical TRUE if valid url, FALSE otherwise
}
\description{
\code{ValidatorUrl} Class for validating website urls
}
\details{
This class provide a methods for validating website urls
}
\section{Methods}{

\describe{
 \item{\code{new()}}{Creates an object of ValidatorUrl class}
 \item{\code{validate(object)}}{Method for validating urls}
}
}

\seealso{
Other Validation Classes: \code{\link{VValidatorCorpusBuilder}},
  \code{\link{Validator0}}, \code{\link{ValidatorClass}},
  \code{\link{ValidatorDate}},
  \code{\link{ValidatorExists}},
  \code{\link{ValidatorLogical}},
  \code{\link{ValidatorNoSpaces}},
  \code{\link{ValidatorNotEmpty}},
  \code{\link{ValidatorNumeric}},
  \code{\link{ValidatorPath}},
  \code{\link{ValidatorString}}, \code{\link{Validator}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
