% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCurator.R
\docType{class}
\name{VCurator}
\alias{VCurator}
\title{VCurator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VCurator
}
\arguments{
\item{className}{Character string containing the name of the class about which the query is rendered.}

\item{objectName}{Character string containing the name of the object about which the query is rendered.}

\item{dateFrom}{An ISO 8601 formatted date from which (inclusive) the data should be returned.}

\item{dateTo}{An ISO 8601 formatted date to which (inclusive) the data should be returned.}

\item{lastN}{A numeric indicating the numeric component of a date or time interval}

\item{unit}{Character string containing the name of the unit associated with teh lastN parameter.  Valid values are "min", "hour", "day", "week", "month"}

\item{child}{The child object}
}
\description{
\code{VCurator} Visitor class responsible for processing queries on
StateManager and Historian class objects.
}
\details{
\strong{VCurator Methods:}
The VCurator methods are as follows:
 \itemize{
  \item{\code{stateManager(object)}}{Method processing queries into StateManager objects.}
  \item{\code{historian(object)}}{Method processing queries into Historian objects.}
}
}
\seealso{
Other Validation Visitor Classes: \code{\link{VValidatorAddChild}},
  \code{\link{VValidatorInit}},
  \code{\link{VValidatorParent}},
  \code{\link{VValidatorRemoveChild}},
  \code{\link{VValidatorRestore}},
  \code{\link{VValidatorSetParent}},
  \code{\link{VValidatorSetQueryTarget}},
  \code{\link{VValidatorSubmitQuery}},
  \code{\link{VValidatorexposeObject}},
  \code{\link{VValidator}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
