% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visitor0.R
\docType{class}
\name{Visitor0}
\alias{Visitor0}
\title{Visitor0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Visitor0
}
\arguments{
\item{lab}{An object of the Lab class.}

\item{collection}{An object of the DocumentCollection class.}

\item{document}{An object of the Document class.}
}
\description{
\code{Visitor0} Interface class which defines a visit operation for each class
of concrete elements within NLPStudio
}
\details{
\strong{Visitor Class Family  Overview:}

The Visitor family of classes is an implementation of the visitor
pattern documented in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). This pattern allows operations
to be defined without changing the clases of the elemnts upon which
it operates.

\strong{Visitor Class Family Participants:}
The participants of the Visitor class family are:
\itemize{
 \item Visitor0: This class which specifies an abstract interface for all concrete visitors.
 \item VisitorOrphan: Visitor class responsible for moving an objects files to the orphan directory when the object has been removed from a lab or a collection.
 \item VisitorPrintObject: Visitor class responsible for printing object meta data to the console.
 }

\strong{Visitor Class Family Collaborators:}
The collaborators of the Visitor0 class are:
\itemize{
 \item Lab: Class for creating and managing documents and workflows.
 \item DocumentCollection: A composite class participant of the Document0 composite class which creates and manages collections of documents
 \item Document: A leaf class participant of the Document0 composite class. Elements include individual documents.
 }

The methods are as follows:
\itemize{
 \item{\code{visitLab(lab)}}{Base method for visiting a object of the Lab class.}
 \item{\code{visitCollection(collection)}}{Base method visiting a object of the DocumentCollection class.}
 \item{\code{visitDocument(document)}}{Base method visiting a object of the Document class.}
}
}
\seealso{
Other Visitor classes: \code{\link{VisitorOrphan}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
