% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VReadState.R
\docType{class}
\name{VReadState}
\alias{VReadState}
\title{VReadState}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VReadState
}
\arguments{
\item{object}{Object to be saved}
}
\description{
\code{VReadState} Visitor class responsible for restoring objects to prior states
}
\details{
\strong{VReadState Methods:}
The VReadState methods are as follows:
 \itemize{
  \item{\code{lab(object)}}{Method for saving the current state of Lab objects.}
  \item{\code{documentCollection(object)}}{Method for saving the current state of DocumentCollection objects.}
  \item{\code{document(object)}}{Method for saving the current state of Document objects.}
}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
