% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VValidator.R
\docType{class}
\name{VValidator}
\alias{VValidator}
\title{VValidator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VValidator
}
\description{
\code{VValidator} Abstract class for the Validation Visitor Classes.
Defines a super-class with certain authorizations that are inherited
by sub-classes.
}
\details{
\strong{VValidator Methods:}
The VValidator methods are as follows:
 \itemize{
  \item{\code{new(object)}}{Not implemented for this abstract class.}

}
}
\seealso{
Other Validation Visitor Classes: \code{\link{VValidatorAddChild}},
  \code{\link{VValidatorCorpusInSource}},
  \code{\link{VValidatorCorpusWebSource}},
  \code{\link{VValidatorInit}},
  \code{\link{VValidatorRemoveChild}},
  \code{\link{VValidatorSetParent}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
