% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VValidatorAddChild.R
\docType{class}
\name{VValidatorAddChild}
\alias{VValidatorAddChild}
\title{VValidatorAddChild}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VValidatorAddChild
}
\arguments{
\item{object}{The parent object}

\item{child}{The child object}
}
\description{
\code{VValidatorAddChild} Visitor class responsible for validating the parameters for the addChild methods
of aggregate and composite classes.
}
\details{
\strong{VValidatorAddChild Methods:}
The VValidatorAddChild methods are as follows:
 \itemize{
  \item{\code{nlpStudio(object, child)}}{Method for validating the addChild method parameters of the NLPStudio object}
  \item{\code{lab(object, child)}}{Method for validating the addChild method parameters of the Lab object}
  \item{\code{documentCollection(object, child)}}{Method for validating the addChild method parameters of the DocumentCollection object.}
  \item{\code{document(object, child)}}{Method for validating the addChild method parameters of the Document object.}
  \item{\code{documentCsv(object, child)}}{Method for validating the addChild method parameters of the DocumentCsv object.}
  \item{\code{documentRdata(object, child)}}{Method for validating the addChild method parameters of the DocumentRdata object.}
  \item{\code{documentXlsx(object, child)}}{Method for validating the addChild method parameters of the DocumentXlsx object.}
}
}
\seealso{
Other Validation Visitor Classes: \code{\link{VValidatorCorpusInSource}},
  \code{\link{VValidatorCorpusWebSource}},
  \code{\link{VValidatorInit}},
  \code{\link{VValidatorRemoveChild}},
  \code{\link{VValidatorSetParent}},
  \code{\link{VValidator}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
