% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VValidatorCorpusBuilder.R
\docType{class}
\name{VValidatorCorpusBuilder}
\alias{VValidatorCorpusBuilder}
\title{VValidatorCorpusBuilder}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VValidatorCorpusBuilder
}
\arguments{
\item{object}{The object in its current state}

\item{...}{Parameters}
}
\description{
\code{VValidatorCorpusBuilder} Visitor class responsible for validating the initialization CorpusBuilder objects
}
\section{VValidatorCorpusBuilder Methods}{

 \describe{
  \item{corpusBuilderRawWeb}{Concrete builder sub-class that produces the Raw Corpus object. }
  \item{corpusBuilderRefined}{Concrete builder sub-class that produces the Refined Corpus object. }
  \item{corpusBuilderReshaped}{Concrete builder sub-class that produces the Reshaped Corpus object. }
  }
}

\seealso{
Other Validation Classes: \code{\link{Validator0}},
  \code{\link{ValidatorClass}},
  \code{\link{ValidatorDate}},
  \code{\link{ValidatorExists}}, \code{\link{ValidatorIO}},
  \code{\link{ValidatorLogical}},
  \code{\link{ValidatorNoSpaces}},
  \code{\link{ValidatorNotEmpty}},
  \code{\link{ValidatorNumeric}},
  \code{\link{ValidatorPath}},
  \code{\link{ValidatorString}},
  \code{\link{ValidatorUrl}}, \code{\link{Validator}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
