% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VValidatorCorpusInSource.R
\docType{class}
\name{VValidatorCorpusInSource}
\alias{VValidatorCorpusInSource}
\title{VValidatorCorpusInSource}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VValidatorCorpusInSource
}
\arguments{
\item{object}{The Corpus object}

\item{source}{Corpus object from which the object will be sourced}
}
\description{
\code{VValidatorCorpusInSource} Visitor class responsible for validating the
parameters for the CorpusInSource method in the Corpus Object.
}
\section{Corpus Core Methods}{

 \describe{
  \item{\code{new()}}{Instantiates this visitor object. }
  \item{\code{object, source}}{The method that performs the validation of the corpus object.}
 }
}

\seealso{
Other Validation Visitor Classes: \code{\link{VValidatorAddChild}},
  \code{\link{VValidatorCorpusWebSource}},
  \code{\link{VValidatorInit}},
  \code{\link{VValidatorRemoveChild}},
  \code{\link{VValidatorSetParent}},
  \code{\link{VValidator}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
