% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VValidatorCorpusSourceWeb.R
\docType{class}
\name{VValidatorCorpusSourceWeb}
\alias{VValidatorCorpusSourceWeb}
\title{VValidatorCorpusSourceWeb}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VValidatorCorpusSourceWeb
}
\arguments{
\item{object}{The Corpus object}

\item{source}{Character string containing url from which the Corpus object will be sourced.}
}
\description{
\code{VValidatorCorpusSourceWeb} Visitor class responsible for validating the
parameters for the CorpusSourceWeb method in the Corpus Object.
}
\section{Corpus Core Methods}{

 \describe{
  \item{\code{new()}}{Instantiates this visitor object. }
  \item{\code{object, source}}{The method that performs the validation of the corpus object.}
 }
}

\seealso{
Other Validation Visitor Classes: \code{\link{VValidatorAddChild}},
  \code{\link{VValidatorCorpusInSource}},
  \code{\link{VValidatorInit}},
  \code{\link{VValidatorRemoveChild}},
  \code{\link{VValidatorSetParent}},
  \code{\link{VValidator}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
