% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VValidatorGetObject.R
\docType{class}
\name{VValidatorexposeObject}
\alias{VValidatorexposeObject}
\title{VValidatorexposeObject}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VValidatorexposeObject
}
\arguments{
\item{object}{The object for which the exposeObject method is invoked.}

\item{requester}{The object invoking the exposeObject method.}
}
\description{
\code{VValidatorexposeObject} Visitor class responsible for validating requests
for object information.  Only permits objects of the State class to invoke
exposeObject methods.
}
\details{
\strong{VValidatorexposeObject Methods:}
The VValidatorexposeObject methods are as follows:
 \itemize{
  \item{\code{nlpStudio(object, ...)}}{Method for validating the instantiation of the NLPStudio object}
  \item{\code{lab(object, ...)}}{Method for validating the instantiation of the Lab object}
  \item{\code{documentCollection(object, ...)}}{Method for validating the instantiation of the DocumentCollection object.}
  \item{\code{documentText(object, ...)}}{Method for validating the instantiation of the DocumentText object.}
  \item{\code{documentCsv(object, ...)}}{Method for validating the instantiation of the DocumentCsv object.}
  \item{\code{documentRdata(object, ...)}}{Method for validating the instantiation of the DocumentRdata object.}
  \item{\code{documentXlsx(object, ...)}}{Method for validating the instantiation of the DocumentXlsx object.}
}
}
\seealso{
Other Validation Visitor Classes: \code{\link{VCurator}},
  \code{\link{VValidatorAddChild}},
  \code{\link{VValidatorInit}},
  \code{\link{VValidatorParent}},
  \code{\link{VValidatorRemoveChild}},
  \code{\link{VValidatorRestore}},
  \code{\link{VValidatorSetParent}},
  \code{\link{VValidatorSetQueryTarget}},
  \code{\link{VValidatorSubmitQuery}},
  \code{\link{VValidator}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
