% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VValidatorInit.R
\docType{class}
\name{VValidatorInit}
\alias{VValidatorInit}
\title{VValidatorInit}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VValidatorInit
}
\arguments{
\item{object}{The object in its current state}

\item{...}{Parameters}
}
\description{
\code{VValidatorInit} Visitor class responsible for validating the initialization objects of all classes
}
\details{
\strong{VValidatorInit Methods:}
The VValidatorInit methods are as follows:
 \itemize{
  \item{\code{nlpStudios(object)}}{Method for validating the instantiation of the NLPStudios object}
  \item{\code{studio(object)}}{Method for validating the instantiation of the Studio object}
  \item{\code{documentCollection(object)}}{Method for validating the instantiation of the DocumentCollection object.}
  \item{\code{document(object)}}{Method for validating the instantiation of the Document object.}
  \item{\code{documentCsv(object)}}{Method for validating the instantiation of the DocumentCsv object.}
  \item{\code{documentRdata(object)}}{Method for validating the instantiation of the DocumentRdata object.}
  \item{\code{documentXlsx(object)}}{Method for validating the instantiation of the DocumentXlsx object.}
}
}
\seealso{
Other Validation Visitor Classes: \code{\link{VValidatorAddChild}},
  \code{\link{VValidatorCorpusInSource}},
  \code{\link{VValidatorCorpusWebSource}},
  \code{\link{VValidatorRemoveChild}},
  \code{\link{VValidatorSetParent}},
  \code{\link{VValidator}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
