% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VValidatorKorpusBuilder.R
\docType{class}
\name{VValidatorKorpusBuilder}
\alias{VValidatorKorpusBuilder}
\title{VValidatorKorpusBuilder}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VValidatorKorpusBuilder
}
\arguments{
\item{object}{The object in its current state}

\item{...}{Parameters}
}
\description{
\code{VValidatorKorpusBuilder} Visitor class responsible for validating the initialization objects of all classes
}
\details{
\strong{VValidatorKorpusBuilder Methods:}
The VValidatorKorpusBuilder methods are as follows:
 \itemize{
  \item{\code{nlpStudio(object)}}{Method for validating the instantiation of the NLPStudio object}
  \item{\code{lab(object)}}{Method for validating the instantiation of the Lab object}
  \item{\code{documentCollection(object)}}{Method for validating the instantiation of the DocumentCollection object.}
  \item{\code{documentText(object)}}{Method for validating the instantiation of the DocumentText object.}
  \item{\code{documentCsv(object)}}{Method for validating the instantiation of the DocumentCsv object.}
  \item{\code{documentRdata(object)}}{Method for validating the instantiation of the DocumentRdata object.}
  \item{\code{documentXlsx(object)}}{Method for validating the instantiation of the DocumentXlsx object.}
}
}
\seealso{
Other Validation Visitor Classes: \code{\link{VValidatorAddChild}},
  \code{\link{VValidatorInit}},
  \code{\link{VValidatorRemoveChild}},
  \code{\link{VValidatorRepair}},
  \code{\link{VValidatorSetParent}},
  \code{\link{VValidator}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
