% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VValidatorRemoveChild.R
\docType{class}
\name{VValidatorRemoveChild}
\alias{VValidatorRemoveChild}
\title{VValidatorRemoveChild}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VValidatorRemoveChild
}
\arguments{
\item{object}{The parent object}

\item{child}{The child object}
}
\description{
\code{VValidatorRemoveChild} Visitor class responsible for validating the parameters for the removeChild methods
of aggregate and composite classes.
}
\details{
\strong{VValidatorRemoveChild Methods:}
The VValidatorRemoveChild methods are as follows:
 \itemize{
  \item{\code{nlpStudio(object, child)}}{Method for validating the removeChild method parameters of the NLPStudio object}
  \item{\code{lab(object, child)}}{Method for validating the removeChild method parameters of the Lab object}
  \item{\code{documentCollection(object, child)}}{Method for validating the removeChild method parameters of the DocumentCollection object.}
  \item{\code{document(object, child)}}{Not implemented for this class.}
  \item{\code{documentCsv(object, child)}}{Not implemented for this class.}
  \item{\code{documentRdata(object, child)}}{Not implemented for this class.}
  \item{\code{documentXlsx(object, child)}}{Not implemented for this class.}
}
}
\seealso{
Other Validation Visitor Classes: \code{\link{VValidatorAddChild}},
  \code{\link{VValidatorCorpusInSource}},
  \code{\link{VValidatorCorpusWebSource}},
  \code{\link{VValidatorInit}},
  \code{\link{VValidatorSetParent}},
  \code{\link{VValidator}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
