% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VValidatorRepair.R
\docType{class}
\name{VValidatorRepair}
\alias{VValidatorRepair}
\title{VValidatorRepair}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VValidatorRepair
}
\arguments{
\item{object}{The document class object}
}
\description{
\code{VValidatorRepair} Visitor class responsible for validating read
requests.
}
\details{
\strong{VValidatorRepair Methods:}
The VValidatorRepair methods are as follows:
 \itemize{
  \item{\code{documentCollection(object)}}{Method for validating a request ro read a DocumentCollection object.}
  \item{\code{documentText(object)}}{Method for validating a request ro read a DocumentText object.}
  \item{\code{documentCsv(object)}}{Method for validating a request ro read a DocumentCsv object.}
  \item{\code{documentRdata(object)}}{Method for validating a request ro read a DocumentRdata object.}
  \item{\code{documentXlsx(object)}}{Method for validating a request ro read a DocumentXlsx object.}
}
}
\seealso{
Other Validation Visitor Classes: \code{\link{VValidatorAddChild}},
  \code{\link{VValidatorInit}},
  \code{\link{VValidatorKorpusBuilder}},
  \code{\link{VValidatorRemoveChild}},
  \code{\link{VValidatorSetParent}},
  \code{\link{VValidator}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
