% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VValidatorRestore.R
\docType{class}
\name{VValidatorRestore}
\alias{VValidatorRestore}
\title{VValidatorRestore}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VValidatorRestore
}
\arguments{
\item{object}{The object in its current state}

\item{...}{Parameters}
}
\description{
\code{VValidatorRestore} Visitor class responsible for validating the
parameters of the restore operation.
}
\details{
\strong{VValidatorRestore Methods:}
The VValidatorRestore methods are as follows:
 \itemize{
  \item{\code{nlpStudio(object, ...)}}{Method for validating the instantiation of the NLPStudio object}
  \item{\code{lab(object, ...)}}{Method for validating the instantiation of the NLPStudio object}
  \item{\code{documentCollection(object, ...)}}{Method for validating the instantiation of the NLPStudio object.}
  \item{\code{documentText(object, ...)}}{Method for validating the instantiation of the NLPStudio object.}
  \item{\code{documentCsv(object, ...)}}{Method for validating the instantiation of the NLPStudio object.}
  \item{\code{documentRdata(object, ...)}}{Method for validating the instantiation of the NLPStudio object.}
  \item{\code{documentXlsx(object, ...)}}{Method for validating the instantiation of the NLPStudio object.}
}
}
\seealso{
Other Validation Visitor Classes: \code{\link{VCurator}},
  \code{\link{VValidatorAddChild}},
  \code{\link{VValidatorInit}},
  \code{\link{VValidatorParent}},
  \code{\link{VValidatorRemoveChild}},
  \code{\link{VValidatorSetParent}},
  \code{\link{VValidatorSetQueryTarget}},
  \code{\link{VValidatorSubmitQuery}},
  \code{\link{VValidatorexposeObject}},
  \code{\link{VValidator}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
