% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VValidatorSetParent.R
\docType{class}
\name{VValidatorSetParent}
\alias{VValidatorSetParent}
\title{VValidatorSetParent}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VValidatorSetParent
}
\arguments{
\item{object}{The target object}

\item{parent}{The parent object}
}
\description{
\code{VValidatorSetParent} Visitor class responsible for validating the parameters for the SetParent methods
of aggregate and composite classes.
}
\details{
\strong{VValidatorSetParent Methods:}
The VValidatorSetParent methods are as follows:
 \itemize{
  \item{\code{lab(object, parent)}}{Method for validating the SetParent method parameters of the Lab object}
  \item{\code{documentCollection(object, parent)}}{Method for validating the SetParent method parameters of the DocumentCollection object.}
  \item{\code{documentText(object, parent)}}{Method for validating the SetParent method parameters of the DocumentText object.}
  \item{\code{documentCsv(object, parent)}}{Method for validating the SetParent method parameters of the DocumentCsv object.}
  \item{\code{documentRdata(object, parent)}}{Method for validating the SetParent method parameters of the DocumentRdata object.}
  \item{\code{documentXlsx(object, parent)}}{Method for validating the SetParent method parameters of the DocumentXlsx object.}
}
}
\seealso{
Other Validation Visitor Classes: \code{\link{VValidatorAddChild}},
  \code{\link{VValidatorInit}},
  \code{\link{VValidatorParent}},
  \code{\link{VValidatorRemoveChild}},
  \code{\link{VValidator}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
