% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VValidatorSetQueryTarget.R
\docType{class}
\name{VValidatorSetQueryTarget}
\alias{VValidatorSetQueryTarget}
\title{VValidatorSetQueryTarget}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VValidatorSetQueryTarget
}
\arguments{
\item{object}{Requesting object}

\item{target}{Target object}
}
\description{
\code{VValidatorSetQueryTarget} Visitor class responsible for validating the request to
SetQueryTarget.
}
\details{
\strong{VValidatorSetQueryTarget Methods:}
The VValidatorSetQueryTarget methods are as follows:
 \itemize{
 \item{\code{nlpStudio(object, target)}}{Verifies that the requesting object can query an NLPStudio object.}
  \item{\code{lab(object, target)}}{Verifies that the requesting object can query an LAB object.}
  \item{\code{documentCollection(object, target)}}{Verifies that the requesting object can query a DocumentCollection object.}
  \item{\code{stateManager(object, target)}}{Verifies that the requesting object can query an StateManager object.}
  \item{\code{historian(object, target)}}{Verifies that the requesting object can query a Historian object.}
}
}
\seealso{
Other Validation Visitor Classes: \code{\link{VCurator}},
  \code{\link{VValidatorAddChild}},
  \code{\link{VValidatorInit}},
  \code{\link{VValidatorParent}},
  \code{\link{VValidatorRemoveChild}},
  \code{\link{VValidatorRestore}},
  \code{\link{VValidatorSetParent}},
  \code{\link{VValidatorSubmitQuery}},
  \code{\link{VValidatorexposeObject}},
  \code{\link{VValidator}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
