% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VValidatorSubmitQuery.R
\docType{class}
\name{VValidatorSubmitQuery}
\alias{VValidatorSubmitQuery}
\title{VValidatorSubmitQuery}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VValidatorSubmitQuery
}
\arguments{
\item{object}{The curator object}
}
\description{
\code{VValidatorSubmitQuery} Visitor class responsible for validating the parameters for the buildQuery methods
of Curator class.
}
\details{
\strong{VValidatorSubmitQuery Methods:}
The VValidatorSubmitQuery methods are as follows:
 \itemize{
  \item{\code{curator(object)}}{Method for validating the buildQuery method parameters of the VCurator object}
}
}
\seealso{
Other Validation Visitor Classes: \code{\link{VCurator}},
  \code{\link{VValidatorAddChild}},
  \code{\link{VValidatorInit}},
  \code{\link{VValidatorParent}},
  \code{\link{VValidatorRemoveChild}},
  \code{\link{VValidatorRestore}},
  \code{\link{VValidatorSetParent}},
  \code{\link{VValidatorSetQueryTarget}},
  \code{\link{VValidatorexposeObject}},
  \code{\link{VValidator}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
