% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VWriter.R
\docType{class}
\name{VWriter}
\alias{VWriter}
\title{VWriter}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VWriter
}
\arguments{
\item{file}{Object of the File family of classes}
}
\description{
\code{VWriter} Visitor class responsible for writing content to files of various formats.
Current formats include .csv, .Rdata, and .txt files
}
\details{
\strong{VWriter Class Overview:}
The VWriter class is an implementation of the visitor design pattern,
as described in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). This visitor pattern allows
new operations to be defined without changing the classes upon which
the visitor method operates.

\strong{VWriter Methods:}
The VWriter class supports csv, Rdata, and text files through the following methods:
 \itemize{
  \item{\code{writeCsv(file)}}{Method for writing csv files.}
  \item{\code{writeRdata(file)}}{Method for writing Rdata files.}
  \item{\code{writeText(file)}}{Method for writing text files.}
}
}
\seealso{
Other Read/Write Classes: \code{\link{Reader}},
  \code{\link{VReader}}, \code{\link{Writer}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
