% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VWriteState.R
\docType{class}
\name{VWriteState}
\alias{VWriteState}
\title{VWriteState}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VWriteState
}
\arguments{
\item{object}{Object to be saved}
}
\description{
\code{VWriteState} Visitor class responsible for saving the current state of the object which accepts this visitor.
}
\details{
\strong{VWriteState Methods:}
The VWriteState methods are as follows:
 \itemize{
  \item{\code{lab(object)}}{Method for saving the current state of Lab objects.}
  \item{\code{documentCollection(object)}}{Method for saving the current state of DocumentCollection objects.}
  \item{\code{document(object)}}{Method for saving the current state of Document objects.}
}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
