% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Write0.R
\docType{class}
\name{Write0}
\alias{Write0}
\title{Write0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Write0
}
\arguments{
\item{content}{A object to be written. Object type will vary by subclass.}

\item{name}{Character string indicating the name of the object to be written}

\item{path}{Character string indicating the location of the file to be written.}
}
\description{
\code{Write0} Class as in abstract class to the WriteBin, WriteText, WriteCsv, WriteRdata write strategies
}
\details{
This class is an abstract class defining the methods to be implemented by
the concrete strategy classes: WriteText, WriteBin, WriteCsv, and WriteRdata.
}
\section{Methods}{

\describe{
 \item{\code{new()}}{Returns an object of the selected concrete write strategy class. This method is overwritten by the concrete write strategy classes}
 \item{\code{writeData(content)}}{Writes the content to the document as per the document object definition. This method is overwritten by the concrete write strategy classes}
}
}

\seealso{
Other Document i/o classes: \code{\link{ReadBin}},
  \code{\link{ReadCsv}}, \code{\link{ReadRdata}},
  \code{\link{ReadText}}, \code{\link{WriteBin}},
  \code{\link{WriteCsv}}, \code{\link{WriteRdata}},
  \code{\link{WriteText}}
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
\keyword{datasets}
