% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteBin.R
\docType{class}
\name{WriteBin}
\alias{WriteBin}
\title{WriteBin}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
WriteBin
}
\arguments{
\item{content}{Character vector of binary content to be written.}

\item{name}{Character string indicating the name of the object to be written,}

\item{path}{Character string indicating the location of the file to be written.}
}
\description{
\code{WriteBin} Class contains the data and methods for writing files in binary format.
}
\details{
This class inherits from the Write0 abstract class and provides the data
and methods for writing files in binary format.
}
\section{Methods}{

\describe{
 \item{\code{writeData(directory, fileName)}}{Writes the content in binary format in the file path indicated in the document object}
}
}

\seealso{
Other Document i/o classes: \code{\link{ReadBin}}
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
\keyword{datasets}
