% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteCsv.R
\docType{class}
\name{WriteCsv}
\alias{WriteCsv}
\title{WriteCsv}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
WriteCsv
}
\arguments{
\item{content}{An object of type csv.}

\item{name}{Character string indicating the name of the object to be written.}

\item{path}{Character string indicating the location of the file to be written.}
}
\description{
\code{WriteCsv} Class contains the data and methods for writing files in csv format.
}
\details{
This class inherits from the Write0 abstract class and provides the data
and methods for writing files in csv format.
}
\section{Methods}{

\describe{
 \item{\code{writeData(directory, fileName)}}{Writes the content in csv format in the file path indicated in the document object}
}
}

\seealso{
Other Document i/o classes: \code{\link{ReadBin}},
  \code{\link{ReadCsv}}, \code{\link{ReadRdata}},
  \code{\link{ReadText}}, \code{\link{Write0}},
  \code{\link{WriteBin}}, \code{\link{WriteRdata}},
  \code{\link{WriteText}}
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
\keyword{datasets}
