% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Writer.R
\docType{class}
\name{Writer}
\alias{Writer}
\title{Writer}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Writer
}
\arguments{
\item{object}{This is an object of the DocumentCollection, Document or File class}
}
\description{
\code{Writer} Class responsible for performing write functionality through
the document and document collection composite hierarchy. This class has
a single method that, upon instantiation, initiates a VWriter object,
passes this visitor object to the document's accept method. The document
accepts the VWriter visitor and the VWriter processes the request or
dispatches another write request on a composite object.
}
\details{
\strong{Writer Class Overview:}
The Writer class is a participant in the VWriter visitor pattern, a design
pattern described in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). This visitor pattern allows
new operations to be defined without changing the classes upon which
the visitor method operates.

\strong{Writer Methods:}
The Writer class has a single method as follows:
 \itemize{
  \item{\code{writeData(object)}}{Method responsible for performing write functionality within the Document composite hierarchy.}
}
}
\seealso{
Other Read/Write Classes: \code{\link{Reader}},
  \code{\link{VReader}}, \code{\link{VWriter}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
