% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteTxt.R
\docType{class}
\name{WriteTxt}
\alias{WriteTxt}
\title{WriteTxt}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
WriteTxt
}
\description{
\code{WriteTxt} Class contains the data and methods for writing files in character format.
}
\details{
This class inherits from the Write0 abstract class and provides the data
and methods for writing files in character format.
}
\section{Methods}{

\describe{
 \item{\code{writeData(directory, fileName)}}{Writes the content in character format in the file path indicated in the document object}
}
}

\seealso{
Other Document i/o classes: \code{\link{ReadBin}},
  \code{\link{ReadCsv}}, \code{\link{ReadRdata}},
  \code{\link{ReadText}}, \code{\link{Write0}},
  \code{\link{WriteBin}}, \code{\link{WriteCsv}},
  \code{\link{WriteRdata}}
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
\keyword{datasets}
