% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Archive.R
\docType{class}
\name{Archive}
\alias{Archive}
\title{Archive}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Archive
}
\arguments{
\item{object}{Object to archived}

\item{name}{Character string containing the name of the archive}

\item{desc}{Character string containing the description of the archive}
}
\description{
\code{Archive} Class responsible for archival and restore of NLPStudio objects.
}
\details{
Class archives NLPStudio objects, including Studio, Data, Corpus & Document
objects. A list of archives is retained and a restoral method enables clients
to restore objects the object to its prior state.
}
\section{Archive core methods}{

 \itemize{
  \item{\code{new(name, object, desc = NULL)}}{Method for instantiating a Archive.}
  \item{\code{getArchives()}}{Method that returns the name of the Archive object.}
  \item{\code{archive()}}{Method for obtaining the Archive file name.}
  \item{\code{restore(archive)}}{Method for obtaining the Archive path.}
 }
}

\section{Archive getter/setter methods}{

 \itemize{
  \item{\code{desc()}}{Method for setting or retrieving the Archive object description.}
 }
}

\section{Other methods}{

 \itemize{
  \item{\code{accept(visitor)}}{Accepts a visitor object.}
  \item{\code{logIt(level = 'Info')}}{Logs events relating to the Archive.}
 }
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
