% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Corpus.R, R/Corpus0.R
\docType{class}
\name{Corpus}
\alias{Corpus}
\alias{Corpus}
\title{Corpus}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Corpus

Corpus
}
\arguments{
\item{name}{A character string containing the name of the Corpus object. This variable is used in the instantiation and remove methods.}

\item{desc}{A chararacter string containing the description of the Corpus.}

\item{document}{An object of one of the Document sub-classes.}

\item{visitor}{An object of one of the visitor classes.}

\item{name}{A character string containing the name of the Corpus object. This variable is used in the instantiation and remove methods.}

\item{desc}{A chararacter string containing the description of the Corpus.}

\item{document}{An object of one of the Document sub-classes.}

\item{visitor}{An object of one of the visitor classes.}
}
\description{
\code{Corpus} Class that defines a corpus or collection of documents

\code{Corpus} Class that defines a corpus or collection of documents
}
\details{
Class contains a collection of text documents along with document
transformations such as NGrams, and POS tagged documents.

Class contains a collection of text documents along with document
transformations such as NGrams, and POS tagged documents.
}
\section{Corpus Core Methods}{

 \describe{
  \item{\code{new(name, desc = NULL, studio = NULL)}}{Creates an object of Corpus Class}
  \item{\code{desc}}{A getter/corpuster method allowing clients to retrieve and corpus the Corpus description variable.}
  \item{\code{parent}}{A getter/corpuster method allowing clients to retrieve and corpus the Studio object to which the Corpus object belongs.}
  \item{\code{getName()}}{Returns the name of the Corpus object.}
  \item{\code{getPath()}}{Returns the path of the Corpus object.}
  \item{\code{getDocuments()}}{Returns the list of Document class objects.}
  \item{\code{addDocument(document)}}{Adds a Document object to the corpus.}
  \item{\code{removeDocument(document)}}{Removes a Document object from the corpus.}
  \item{\code{logIt(level = 'Info', fieldName = NA)}}{Formats the log and calls the LogR class to log an event.}
  \item{\code{accept(visitor)}}{Accepts an object of the Visitor family of classes.}
 }


 \describe{
  \item{\code{new(name, desc = NULL, studio = NULL)}}{Creates an object of Corpus Class}
  \item{\code{desc}}{A getter/corpuster method allowing clients to retrieve and corpus the Corpus description variable.}
  \item{\code{parent}}{A getter/corpuster method allowing clients to retrieve and corpus the Studio object to which the Corpus object belongs.}
  \item{\code{getName()}}{Returns the name of the Corpus object.}
  \item{\code{getPath()}}{Returns the path of the Corpus object.}
  \item{\code{getDocuments()}}{Returns the list of Document class objects.}
  \item{\code{addDocument(document)}}{Adds a Document object to the corpus.}
  \item{\code{removeDocument(document)}}{Removes a Document object from the corpus.}
  \item{\code{logIt(level = 'Info', fieldName = NA)}}{Formats the log and calls the LogR class to log an event.}
  \item{\code{accept(visitor)}}{Accepts an object of the Visitor family of classes.}
 }
}

\author{
John James, \email{jjames@datasciencesalon.org}

John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
