% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Corpus.R
\docType{class}
\name{Corpus}
\alias{Corpus}
\title{Corpus}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Corpus
}
\arguments{
\item{name}{A character string containing the name of the Corpus object. This variable is used in the instantiation and remove methods.}

\item{desc}{A chararacter string containing the description of the Corpus.}

\item{set}{An object of the Set class.}

\item{visitor}{An object of one of the visitor classes.}
}
\description{
\code{Corpus} Class that defines a corpus or collection of data sets
}
\details{
Class defines the Corpus object as a collection of data sets or Set objects.
}
\section{Corpus Core Methods}{

 \describe{
  \item{\code{new(name, desc = NULL, lab = NULL)}}{Creates an object of Corpus Class}
  \item{\code{desc}}{A getter/setter method allowing clients to retrieve and set the Corpus description variable.}
  \item{\code{parent}}{A getter/setter method allowing clients to retrieve and set the Lab object to which the Corpus object belongs.}
  \item{\code{getName()}}{Returns the name of the Corpus object.}
  \item{\code{getPath()}}{Returns the path of the Corpus object.}
  \item{\code{move(parent)}}{Moves the Corpus object to the new parent.}
  \item{\code{getSets()}}{Returns list of set objects.}
  \item{\code{addSet(set)}}{Adds a Set object to the Corpus object.}
  \item{\code{removeSet(set)}}{Removes a Set object from the Corpus object.}
  \item{\code{logIt(level = 'Info', fieldName = NA)}}{Formats the log and calls the LogR class to log an event.}
  \item{\code{accept(visitor)}}{Accepts an object of the Visitor family of classes.}
 }
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
