% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Corpus.R
\docType{class}
\name{Corpus}
\alias{Corpus}
\title{Corpus}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Corpus
}
\arguments{
\item{collection}{FileCollection object.}

\item{field}{Character string name for a field to be added to the Document or Corpus object meta data.}

\item{name}{A character string containing the name of the Corpus object. This variable is used in the instantiation and remove methods.}

\item{visitor}{An object of one of the visitor classes.}
}
\description{
\code{Corpus} Class that defines a corpus or collection of documents
}
\details{
Class contains a collection of text documents along with document
transformations such as NGrams, and POS tagged documents.
}
\section{Corpus Core Methods}{

 \describe{
  \item{\code{new(name, path)}}{Creates an object of Corpus Class}
  \item{\code{getName()}}{Returns the name of the Corpus object.}
  \item{\code{getPath()}}{Returns the path of the Corpus object.}
 }
}

\section{Corpus Sourcing Methods}{

 \describe{
  \item{\code{download(url, name)}}{Creates an object of Corpus Class}
  \item{\code{zipFile()}}{Returns the name of the Corpus object.}
  \item{\code{unZipFile()}}{Returns the path of the Corpus object.}
 }
}

\section{IO Methods}{

 \describe{
  \item{\code{read(io = NULL)}}{Reads a corpus into the Corpus object.}
  \item{\code{write(io = NULL)}}{Writes a Corpus object to file.}
 }
}

\section{Analysis Methods}{

 \describe{
  \item{\code{stats()}}{Produces a data frame with basic descriptive statistics for the corpus. }
  \item{\code{diversity}}{Produces a data frame of lexical diversity measures for the corpus.}
  \item{\code{readability}}{Produces a data frame of readability measures for the corpus.}
 }
}

\section{Meta Data Methods}{

 \describe{
  \item{\code{docMeta(field)}}{Creates a document meta data field.}
  \item{\code{corpusMeta(field)}}{Creates a corpus meta data field.}
 }
}

\section{Other Methods}{

 \describe{
  \item{\code{logIt(level = 'Info', fieldName = NA)}}{Formats the log and calls the LogR class to log an event.}
  \item{\code{accept(visitor)}}{Accepts an object of the Visitor family of classes.}
 }
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
