% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorpusDirector.R, R/KorpusDirector.R
\docType{class}
\name{CorpusDirector}
\alias{CorpusDirector}
\alias{CorpusDirector}
\title{CorpusDirector}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CorpusDirector

CorpusDirector
}
\arguments{
\item{builder}{Builder object for the Corpus object}

\item{size}{Numeric proportion of total object size to retain for training. Validation and test sets retain 20 percent of the corpus each.}

\item{parserCmd}{Parser command object for parsing the data during Processing}

\item{builder}{Corpus builder object}
}
\description{
\code{CorpusDirector} Class that directs the construction of various types of corpora.

\code{CorpusDirector} Class that directs the building of Corpus objects
}
\details{
Class directs the construction of the raw, refined, reshaped and processed corpora

The Document family of classes is an implementation of the builder
pattern documented in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). This pattern allows the
construction process to be defined at runtime.

The Document family of classes is an implementation of the builder
pattern documented in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). This pattern allows the
construction process to be defined at runtime.
}
\section{Corpus Builder Participants}{

 \describe{
  \item{CorpusBuilder0}{The abstract builder interface. Defines the interface for concrete corpus builder sub-classes. }
  \item{CorpusBuilderRawWeb}{Concrete builder subclass that produces the Raw Corpus object. }
  \item{CorpusBuilderRefined}{Concrete builder subclass that produces the Refined Corpus object. }
  \item{CorpusBuilderReshaped}{Concrete builder subclass that produces the Reshaped Corpus object. }
  \item{CorpusBuilderProcessed}{Concrete builder subclass that produces the Processed Corpus object. }
  \item{CorpusDirector}{Class that builds the corpus through the concrete builder interfaces.}
  \item{Corpus}{The corpus product.}
  }
}

\section{CorpusDirector Methods}{

 \describe{
  \item{\code{new(builder, parserCmd = NULL)}}{Instantiates the director object}
  \item{\code{construct()}}{Initiates the construction process for the builder object.}
 }
}

\section{Corpus Director Methods}{

 \describe{
  \item{\code{new(builder)}}{Creates an object of Corpus Class}
  }
}

\author{
John James, \email{jjames@datasciencesalon.org}

John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
