% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDirector.R, R/KorpusDirector.R
\docType{class}
\name{CorpusDirector}
\alias{CorpusDirector}
\alias{CorpusDirector}
\title{CorpusDirector}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CorpusDirector

CorpusDirector
}
\arguments{
\item{builder}{Builder object for the Corpus object}

\item{builder}{Corpus builder object}
}
\description{
\code{CorpusDirector} Class that directs the construction of various types of corpora.

\code{CorpusDirector} Class that directs the building of Corpus objects
}
\details{
Class directs the construction of the raw, refined, reshaped and processed corpora

The Document family of classes is an implementation of the builder
pattern documented in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). This pattern allows the
construction process to be defined at runtime.

The Document family of classes is an implementation of the builder
pattern documented in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). This pattern allows the
construction process to be defined at runtime.
}
\section{Corpus Builder Participants}{

 \describe{
  \item{CorpusBuilder0}{This abstract builder interface. Defines the methods for concrete corpus builder sub-classes. }
  \item{CorpusBuilderCv}{Concrete builder sub-class that produces the a single training, validation, and test corpora. }
  \item{CorpusBuilderKFoldCv}{Concrete builder sub-class that produces the K-Fold cross validation corpora. }
  \item{CorpusDirector}{Class that builds the corpus through the concrete builder classes..}
  \item{Corpus}{The corpus product.}
  }
}

\section{CorpusDirector Methods}{

 \describe{
  \item{\code{new(builder)}}{Instantiates the director object}
  \item{\code{construct()}}{Initiates the construction process for the builder object.}
 }
}

\section{Corpus Director Methods}{

 \describe{
  \item{\code{new(builder)}}{Creates an object of Corpus Class}
  }
}

\author{
John James, \email{jjames@datasciencesalon.org}

John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
