% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KorpusDirector.R
\docType{class}
\name{CorpusDirector}
\alias{CorpusDirector}
\title{CorpusDirector}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CorpusDirector
}
\arguments{
\item{builder}{Corpus builder object}
}
\description{
\code{CorpusDirector} Class that directs the building of Corpus objects
}
\details{
The Document family of classes is an implementation of the builder
pattern documented in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). This pattern allows the
construction process to be defined at runtime.
}
\section{Corpus Director Methods}{

 \describe{
  \item{\code{new(builder)}}{Creates an object of Corpus Class}
  }
}

\seealso{
Other Corpus build family: \code{\link{Builder0}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
