% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document.R
\docType{class}
\name{Document}
\alias{Document}
\title{Document}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Document
}
\arguments{
\item{name}{Character string indicating the name of the document or file. Required for all objects.}

\item{desc}{Character string containing the description of the document.}

\item{content}{Nested list of content to be written to files.}

\item{fileName}{Character string indicating File object's file name.}

\item{korpus}{A Korpus object}

\item{reference}{A list of data frames containing reference information used in the cleaning methods}

\item{nGrams}{Numeric indicator of the number of nGrams to produce}
}
\description{
\code{Document} An abstract class that defines the interfaces for the document sub-classes
}
\details{
The Document family of classes is an implementation of the strategy
pattern documented in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). This pattern allows behavior
of a class to be defined at runtime.
}
\section{Document family of classes participants}{

The participants of the Document class are:
\itemize{
 \item Document: This "abstract leaf" class defines the interface for
 DocumentText, DocumentRdata, DocumentCsv, and DocumentXlsx sub-classes.
 \item DocumentText: This "concrete leaf" class for text documents.
 \item DocumentCsv: This "concrete leaf" class for csv documents.
 \item DocumentRdata: This "concrete leaf" class for RData documents.
 \item DocumentXlsx: This "concrete leaf" class for excel documents.
 }
}

\section{Document class collaborators}{

The collaborators of the Document family  are:
 \itemize{
  \item Korpus: Class containing one or more documents
  \item Reader: Class responsible for initiating the document read operation.
  \item Writer: Class responsible for initiating the document write operation.
  \item FileManager: Class responsible for managing and repairing files
  \item VReader: Visitor class responsible for performing read operations through the Document hierarchy.
  \item VWriter: Visitor class responsible for performing write operations through the Document hierarchy.
  \item NGrams: Class responsible for creating nGram representations of the documents
  \item POSTags: Class responsible for creating POS tagged representations of the documents
 }

 @section Document methods:
 \itemize{
  \item{\code{new(name, fileName, desc)}}{Method for instantiating a document}
  \item{\code{getName()}}{Method for obtaining the document name.}
  \item{\code{desc()}}{Method for retrieving and setting the document description.}
  \item{\code{korpus()}}{Method for retrieving and setting the korpus for the document.}
  \item{\code{readDocument()}}{Method for initiating the read operation for a document.}
  \item{\code{writeDocument(content)}}{Method for initiating the write operation for a document.}
  \item{\code{repairDocument()}}{Method for initiating the repair operation on a document.}
  \item{\code{cleanDocument(reference)}}{Method for cleaning the document.}
  \item{\code{nGramDocument(nGrams)}}{Method for creating an nGram representation of the document.}
  \item{\code{posTagDocument(nGrams)}}{Method for creating an POS tagged representation of the document.}
 }
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
