% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document0.R
\docType{class}
\name{Document0}
\alias{Document0}
\title{Document0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Document0
}
\arguments{
\item{filePath}{Character string indicating the file path for a document}

\item{parent}{Object of the Corpus or Set classes to which this document belongs}

\item{io}{An object of one of the IO classes used for reading and writing.}
}
\description{
\code{Document0} Abstract clasS for the Document Strategy Classes
}
\details{
Defines the base methods for the two document sub-classes, the corpus
document and the cross validation set document.
}
\section{Document0 core methods}{

 \itemize{
  \item{\code{new()}}{Method for instantiating a document. Not implemented for the abstract class.}
  \item{\code{getName()}}{Method that returns the name of the Document object. }
  \item{\code{getFileName()}}{Method for obtaining the document file name. .}
  \item{\code{getPath()}}{Method for obtaining the document path. }
  \item{\code{getContent()}}{Method for obtaining the document content.}
 }
}

\section{Document0 getter/setter methods}{

 \itemize{
  \item{\code{desc()}}{Method for setting or retrieving the Document0 object description.}
 }

 @section Document0 IO methods:
 \itemize{
  \item{\code{addContent(content)}}{Method for adding content to a document.}
  \item{\code{read(io)}}{Method for reading a document. }
  \item{\code{write(io)}}{Method for writing a document. }
 }
}

\section{Document0 aggregation method}{

 \itemize{
  \item{\code{move(parent)}}{Moves a document to the designated parent corpus. }
 }
}

\section{Other methods}{

 \itemize{
  \item{\code{accept(visitor)}}{Accepts a visitor object. Not implemented for this abstract class}
  \item{\code{logIt(level = 'Info')}}{Logs events relating to the document.}
 }
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
