% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document0.R
\docType{class}
\name{Document0}
\alias{Document0}
\title{Document0}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Document0
}
\arguments{
\item{desc}{Character string containing the document description.}

\item{filePath}{Character string indicating the file path for a document.}

\item{io}{An object of one of the IO classes used for reading and writing.}
}
\description{
\code{Document0} Abstract class that defines the interface for the document sub-classes
}
\details{
This abstract class defines the interface for the family of document classes
responsible for creating, manipulating, processing, and transforming
documents.
}
\section{Document Family Participants}{

 \itemize{
  \item{Document0}{This abstract class defines the interface for the other classes.}
  \item{DocumentText}{Class for text documents.}
  \item{DocumentNGrams}{Class for NGrams.}
  \item{DocumentPOSTags}{Class for POS tags.}
  }
}

\section{Document methods}{

 \itemize{
  \item{\code{new(filePath, desc = NULL)}}{Method for instantiating a document.}
  \item{\code{name()}}{Method for setting or retrieving the Document object name.}
  \item{\code{desc()}}{Method for setting or retrieving the Document object description.}
  \item{\code{getName()}}{Method that returns the name of the Document object.}
  \item{\code{getFileName()}}{Method for obtaining the document file name.}
  \item{\code{getPath()}}{Method for obtaining the document path.}
  \item{\code{readDocument(io)}}{Method for initiating the read operation for a document.}
  \item{\code{writeDocument(io)}}{Method for initiating the write operation for a document.}
 }
}

\seealso{
Other Document classes: \code{\link{DocumentText}},
  \code{\link{Preprocess}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
