% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOBin.R
\docType{class}
\name{IOBin}
\alias{IOBin}
\title{IOBin}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
IOBin
}
\arguments{
\item{document}{Object of the Document family of classes}
}
\description{
\code{IOBin} Class responsible for reading and writing binary files.
}
\details{
\strong{IO Class Overview:}
The IOBin class is an implementation of the strategy design pattern,
as described in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). This strategy pattern accommodates
various formats and allows the behavior to be defined / selected at run time.

\strong{IOBin Methods:}
The IOBin class supports csv, Rdata, and text files through the following methods:
 \itemize{
  \item{\code{read(document)}}{Read method.}
  \item{\code{write(document)}}{Write method.}
}
}
\seealso{
Other Input / Output Classes: \code{\link{IO0}},
  \code{\link{IOCSV}}, \code{\link{IORdata}},
  \code{\link{IORds}}, \code{\link{IOText}},
  \code{\link{IOXlsx}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
