% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IORds.R
\docType{class}
\name{IORds}
\alias{IORds}
\title{IORds}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
IORds
}
\arguments{
\item{document}{Object of the Document family of classes}
}
\description{
\code{IORds} Class responsible for reading and writing rdata documents.
}
\details{
\strong{IO Class Overview:}
The IORds class is an implementation of the strategy design pattern,
as described in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). This strategy pattern accommodates
various formats and allows the behavior to be defined / selected at run time.

\strong{IORds Methods:}
The IORds class supports csv, Rds, and text files through the following methods:
 \itemize{
  \item{\code{read(document)}}{Read method.}
  \item{\code{write(document)}}{Write method.}
}
}
\seealso{
Other Input / Output Classes: \code{\link{IO0}},
  \code{\link{IOBin}}, \code{\link{IOCSV}},
  \code{\link{IORdata}}, \code{\link{IOText}},
  \code{\link{IOXlsx}}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
