% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Korpus.R
\docType{class}
\name{Korpus}
\alias{Korpus}
\title{Korpus}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Korpus
}
\arguments{
\item{name}{A character string containing the name of the Korpus object. This variable is used in the instantiation and remove methods.}

\item{desc}{A chararacter string containing the description of the Korpus.}

\item{document}{An object of one of the Document sub-classes.}

\item{visitor}{An object of one of the visitor classes.}
}
\description{
\code{Korpus} Class that defines a corpus or collection of documents
}
\details{
The class one or several documents in text form, as well as various
transformations such as nGrams and POS tags
}
\section{Korpus Core Methods}{

 \describe{
  \item{\code{new(name, desc = NULL, lab = NULL)}}{Creates an object of Korpus Class}
  \item{\code{desc}}{A getter/setter method allowing clients to retrieve and set the Korpus description variable.}
  \item{\code{lab}}{A getter/setter method allowing clients to retrieve and set the Lab object to which the Korpus object belongs.}
  \item{\code{getName()}}{Returns the name of the Korpus object.}
  \item{\code{getPath()}}{Returns the path of the Korpus object.}
  \item{\code{logIt(level = 'Info', fieldName = NA)}}{Formats the log and calls the LogR class to log an event.}
  \item{\code{accept(visitor)}}{Accepts an object of the Visitor family of classes.}
 }
}

\section{Korpus Processing Methods}{

 \describe{
  \item{\code{obtainKorpus()}}{Method for initiating the repair operation on a document.}
  \item{\code{repairKorpus()}}{Method for initiating the repair operation on a document.}
  \item{\code{splitCorpus()}}{Method for initiating the repair operation on a document.}
}
}

\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
