% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab.R
\docType{class}
\name{Lab}
\alias{Lab}
\title{Lab}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Lab
}
\arguments{
\item{name}{A character string containing the name of the Lab object. This variable is used in the instantiation and remove methods.}

\item{desc}{A chararacter string containing the description of the Lab}

\item{model}{An object of the Model class}

\item{visitor}{An object of one of the visitor classes.}
}
\description{
\code{Lab} Class in which models are created, executed and evaluated.
}
\details{
The environment in which NLP happens. There are two groups of methods. The
core methods allow clients to instantiate labs and to obtain their basic
information.  Aggregate methods enable clients to add, retrieve and remove
models from the lab.

\strong{Lab Core Methods:}
 \describe{
  \item{\code{new(name, desc = NULL)}}{Creates an object of Lab Class}
  \item{\code{desc}}{A getter/setter method allowing clients to retrieve and set the Lab description variable.}
  \item{\code{getName()}}{Returns the name of the Lab object.}
  \item{\code{getPath()}}{Returns the path of the Lab object.}
  \item{\code{getLogs()}}{Returns the LogR object for the Lab object.}
  \item{\code{logIt(level = 'Info', fieldName = NA)}}{Formats the log and calls the LogR class to log an event.}
 }

\strong{Lab Aggregate Methods:}
 \describe{
  \item{\code{getModels()}}{Retrieves the list of models for the lab.}
  \item{\code{addModel(model)}}{Adds a model to the Lab object.}
  \item{\code{removeModel(model)}}{Removes a model from the Lab object. The model is archived in the NLPStudio archives.}
}

\strong{Lab Visitor Methods:}
 \describe{
  \item{\code{accept(visitor)}}{Accepts an object of the Visitor family of classes.}
}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
