% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLPStudio.R
\docType{class}
\name{NLPStudio}
\alias{NLPStudio}
\title{NLPStudio}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NLPStudio
}
\description{
\code{NLPStudio} Containing class containing NLP data and modeling pipelines.
}
\details{
This class creates and manages data science pipelines. Multiple pipelines
can be created, with separate data sets, feature engineering, and models.

\strong{NLPStudio Core Methods:}
\describe{
 \item{\code{new()}}{Initializes the NLPStudio as s singleton object at load time.}

\item{\code{getInstance()}}{Returns the current NLPStudio instance object. This will be the only instantiation called "nlpStudio.},
}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
