% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validator.R
\docType{class}
\name{Validator}
\alias{Validator}
\title{Validator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Validator
}
\arguments{
\item{object}{Originator object}

\item{target}{Object which is the target of a query. Used by the Curator methods}

\item{content}{Character vector containing content to be written by a write originator.}
}
\value{
A logical, TRUE if validation criteria are met, FALSE, otherwise.
}
\description{
\code{Validator} Class responsible for validation of requests pertaining to:
\itemize{
 \item Object instantiation
 \item Composition and Aggregation: Requests to manipulate aggregate and composite objects
 \item History: Requests to read historical records for an object or objects.#'
 \item State: Requests to read, save, and restore object states
 \item Document I/O: Requests to read and write documents.
}
}
\section{Validator methods}{

This section summarizes the methods in the Validator class.

\strong{Object Instantiation Methods::}
\describe{
 \item{\code{new()}}{Creates an object of Validator class}
 \item{\code{init(object)}}{Dispatches the initialization validation visitor, via the accept method of object.}
 }

\strong{Composition and Aggregation Methods:}
\describe{
 \item{\code{addChild(object)}}{Dispatches the addChild validation visitor, via the accept method of object.}
 \item{\code{removeChild(object)}}{Dispatches the removeChild validation visitor, via the accept method of object.}
 \item{\code{setParent(object, parent)}}{Dispatches the setParent validation visitor, via the accept method of object.}
 }

\strong{Document I/O Methods:}
\describe{
 \item{\code{read(object)}}{Dispatches the read validation visitor, via the accept method of object.}
 \item{\code{write(object)}}{Dispatches the write validation visitor, via the accept method of object.}
}
}

\section{Validation Class Parameters}{
NA
}

\seealso{
Other Validation Classes: \code{\link{Validator0}},
  \code{\link{ValidatorClass}},
  \code{\link{ValidatorDate}},
  \code{\link{ValidatorExists}},
  \code{\link{ValidatorLogical}},
  \code{\link{ValidatorNoSpaces}},
  \code{\link{ValidatorNotEmpty}},
  \code{\link{ValidatorNumeric}},
  \code{\link{ValidatorPath}},
  \code{\link{ValidatorString}}, \code{\link{ValidatorUrl}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\keyword{datasets}
