% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VReader.R
\docType{class}
\name{VReader}
\alias{VReader}
\title{VReader}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VReader
}
\arguments{
\item{file}{Object of the File family of classes}
}
\description{
\code{VReader} Visitor class responsible for reading documents of various formats. Current formats include .csv, .Rdata, and .txt files
}
\details{
\strong{VReader Class Overview:}
The VReader class is an implementation of the visitor design pattern,
as described in the book "Design Patterns: Elements of Reusable
Object-Oriented Software" by Erich Gamma, Richard Helm, Ralph Johnson
and John Vlissides (hence Gang of Four). This visitor pattern allows
new operations to be defined without changing the classes upon which
the visitor method operates.

\strong{VReader Methods:}
The VReader class supports csv, Rdata, and text files through the following methods:
 \itemize{
  \item{\code{readCsv(file)}}{Method for reading csv files.}
  \item{\code{readRdata(file)}}{Method for reading Rdata files.}
  \item{\code{readText(file)}}{Method for reading text files.}
}
}
\author{
John James, \email{jjames@DataScienceSalon.org}
}
\keyword{datasets}
